Split whole package in subpackages
