+++ pyrism/core/__init__.py
@@ -1,3 +1,3 @@
-from ._core import Kernel, Scattering
+from ._core import Kernel
 from .auxiliary import (ReflectanceResult, EmissivityResult, SailResult, BRF, BSC, BRDF, dB, sec,
                         cot, rad, align_all, load_param, linear)
