+++ pyrism/data/light_spectra.txt
