+++ pyrism/data/soil_reflectance.txt
