--- pyrism/models/__init__.py
@@ -1,8 +1,0 @@
-from .library import get_data_one, get_data_two
-from .models import (VolScatt, LIDF, PROSPECT, Mie, DielConstant, CorrFunc, exponential, gaussian, xpower,
-                     I2EM, LSM, SAIL)
-
-try:
-    lib = get_data_two()
-except IOError:
-    lib = get_data_one()
