+++ pyrism/scattering/__init__.py
@@ -1,1 +1,2 @@
 from rayleigh import Rayleigh
+from mie import Mie
