+++ pyrism/scattering/rayleigh/core/rayleigh_phase_c.c
@@ -832,7 +832,7 @@ typedef struct {
 } __Pyx_BufFmt_Context;
 
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":730
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":730
  * # in Cython to enable them only on the right systems.
  * 
  * ctypedef npy_int8       int8_t             # <<<<<<<<<<<<<<
@@ -841,7 +841,7 @@ typedef struct {
  */
 typedef npy_int8 __pyx_t_5numpy_int8_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":731
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":731
  * 
  * ctypedef npy_int8       int8_t
  * ctypedef npy_int16      int16_t             # <<<<<<<<<<<<<<
@@ -850,7 +850,7 @@ typedef npy_int8 __pyx_t_5numpy_int8_t;
  */
 typedef npy_int16 __pyx_t_5numpy_int16_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":732
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":732
  * ctypedef npy_int8       int8_t
  * ctypedef npy_int16      int16_t
  * ctypedef npy_int32      int32_t             # <<<<<<<<<<<<<<
@@ -859,7 +859,7 @@ typedef npy_int16 __pyx_t_5numpy_int16_t;
  */
 typedef npy_int32 __pyx_t_5numpy_int32_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":733
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":733
  * ctypedef npy_int16      int16_t
  * ctypedef npy_int32      int32_t
  * ctypedef npy_int64      int64_t             # <<<<<<<<<<<<<<
@@ -868,7 +868,7 @@ typedef npy_int32 __pyx_t_5numpy_int32_t;
  */
 typedef npy_int64 __pyx_t_5numpy_int64_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":737
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":737
  * #ctypedef npy_int128     int128_t
  * 
  * ctypedef npy_uint8      uint8_t             # <<<<<<<<<<<<<<
@@ -877,7 +877,7 @@ typedef npy_int64 __pyx_t_5numpy_int64_t;
  */
 typedef npy_uint8 __pyx_t_5numpy_uint8_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":738
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":738
  * 
  * ctypedef npy_uint8      uint8_t
  * ctypedef npy_uint16     uint16_t             # <<<<<<<<<<<<<<
@@ -886,7 +886,7 @@ typedef npy_uint8 __pyx_t_5numpy_uint8_t;
  */
 typedef npy_uint16 __pyx_t_5numpy_uint16_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":739
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":739
  * ctypedef npy_uint8      uint8_t
  * ctypedef npy_uint16     uint16_t
  * ctypedef npy_uint32     uint32_t             # <<<<<<<<<<<<<<
@@ -895,7 +895,7 @@ typedef npy_uint16 __pyx_t_5numpy_uint16_t;
  */
 typedef npy_uint32 __pyx_t_5numpy_uint32_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":740
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":740
  * ctypedef npy_uint16     uint16_t
  * ctypedef npy_uint32     uint32_t
  * ctypedef npy_uint64     uint64_t             # <<<<<<<<<<<<<<
@@ -904,7 +904,7 @@ typedef npy_uint32 __pyx_t_5numpy_uint32_t;
  */
 typedef npy_uint64 __pyx_t_5numpy_uint64_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":744
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":744
  * #ctypedef npy_uint128    uint128_t
  * 
  * ctypedef npy_float32    float32_t             # <<<<<<<<<<<<<<
@@ -913,7 +913,7 @@ typedef npy_uint64 __pyx_t_5numpy_uint64_t;
  */
 typedef npy_float32 __pyx_t_5numpy_float32_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":745
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":745
  * 
  * ctypedef npy_float32    float32_t
  * ctypedef npy_float64    float64_t             # <<<<<<<<<<<<<<
@@ -922,7 +922,7 @@ typedef npy_float32 __pyx_t_5numpy_float32_t;
  */
 typedef npy_float64 __pyx_t_5numpy_float64_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":754
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":754
  * # The int types are mapped a bit surprising --
  * # numpy.int corresponds to 'l' and numpy.long to 'q'
  * ctypedef npy_long       int_t             # <<<<<<<<<<<<<<
@@ -931,7 +931,7 @@ typedef npy_float64 __pyx_t_5numpy_float64_t;
  */
 typedef npy_long __pyx_t_5numpy_int_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":755
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":755
  * # numpy.int corresponds to 'l' and numpy.long to 'q'
  * ctypedef npy_long       int_t
  * ctypedef npy_longlong   long_t             # <<<<<<<<<<<<<<
@@ -940,7 +940,7 @@ typedef npy_long __pyx_t_5numpy_int_t;
  */
 typedef npy_longlong __pyx_t_5numpy_long_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":756
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":756
  * ctypedef npy_long       int_t
  * ctypedef npy_longlong   long_t
  * ctypedef npy_longlong   longlong_t             # <<<<<<<<<<<<<<
@@ -949,7 +949,7 @@ typedef npy_longlong __pyx_t_5numpy_long_t;
  */
 typedef npy_longlong __pyx_t_5numpy_longlong_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":758
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":758
  * ctypedef npy_longlong   longlong_t
  * 
  * ctypedef npy_ulong      uint_t             # <<<<<<<<<<<<<<
@@ -958,7 +958,7 @@ typedef npy_longlong __pyx_t_5numpy_longlong_t;
  */
 typedef npy_ulong __pyx_t_5numpy_uint_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":759
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":759
  * 
  * ctypedef npy_ulong      uint_t
  * ctypedef npy_ulonglong  ulong_t             # <<<<<<<<<<<<<<
@@ -967,7 +967,7 @@ typedef npy_ulong __pyx_t_5numpy_uint_t;
  */
 typedef npy_ulonglong __pyx_t_5numpy_ulong_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":760
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":760
  * ctypedef npy_ulong      uint_t
  * ctypedef npy_ulonglong  ulong_t
  * ctypedef npy_ulonglong  ulonglong_t             # <<<<<<<<<<<<<<
@@ -976,7 +976,7 @@ typedef npy_ulonglong __pyx_t_5numpy_ulong_t;
  */
 typedef npy_ulonglong __pyx_t_5numpy_ulonglong_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":762
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":762
  * ctypedef npy_ulonglong  ulonglong_t
  * 
  * ctypedef npy_intp       intp_t             # <<<<<<<<<<<<<<
@@ -985,7 +985,7 @@ typedef npy_ulonglong __pyx_t_5numpy_ulonglong_t;
  */
 typedef npy_intp __pyx_t_5numpy_intp_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":763
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":763
  * 
  * ctypedef npy_intp       intp_t
  * ctypedef npy_uintp      uintp_t             # <<<<<<<<<<<<<<
@@ -994,7 +994,7 @@ typedef npy_intp __pyx_t_5numpy_intp_t;
  */
 typedef npy_uintp __pyx_t_5numpy_uintp_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":765
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":765
  * ctypedef npy_uintp      uintp_t
  * 
  * ctypedef npy_double     float_t             # <<<<<<<<<<<<<<
@@ -1003,7 +1003,7 @@ typedef npy_uintp __pyx_t_5numpy_uintp_t;
  */
 typedef npy_double __pyx_t_5numpy_float_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":766
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":766
  * 
  * ctypedef npy_double     float_t
  * ctypedef npy_double     double_t             # <<<<<<<<<<<<<<
@@ -1012,7 +1012,7 @@ typedef npy_double __pyx_t_5numpy_float_t;
  */
 typedef npy_double __pyx_t_5numpy_double_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":767
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":767
  * ctypedef npy_double     float_t
  * ctypedef npy_double     double_t
  * ctypedef npy_longdouble longdouble_t             # <<<<<<<<<<<<<<
@@ -1021,7 +1021,7 @@ typedef npy_double __pyx_t_5numpy_double_t;
  */
 typedef npy_longdouble __pyx_t_5numpy_longdouble_t;
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":10
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":9
  * 
  * DTYPE = np.float
  * ctypedef np.float_t DTYPE_t             # <<<<<<<<<<<<<<
@@ -1056,7 +1056,7 @@ static CYTHON_INLINE __pyx_t_double_complex __pyx_t_double_complex_from_parts(do
 
 /*--- Type declarations ---*/
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":769
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":769
  * ctypedef npy_longdouble longdouble_t
  * 
  * ctypedef npy_cfloat      cfloat_t             # <<<<<<<<<<<<<<
@@ -1065,7 +1065,7 @@ static CYTHON_INLINE __pyx_t_double_complex __pyx_t_double_complex_from_parts(do
  */
 typedef npy_cfloat __pyx_t_5numpy_cfloat_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":770
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":770
  * 
  * ctypedef npy_cfloat      cfloat_t
  * ctypedef npy_cdouble     cdouble_t             # <<<<<<<<<<<<<<
@@ -1074,7 +1074,7 @@ typedef npy_cfloat __pyx_t_5numpy_cfloat_t;
  */
 typedef npy_cdouble __pyx_t_5numpy_cdouble_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":771
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":771
  * ctypedef npy_cfloat      cfloat_t
  * ctypedef npy_cdouble     cdouble_t
  * ctypedef npy_clongdouble clongdouble_t             # <<<<<<<<<<<<<<
@@ -1083,7 +1083,7 @@ typedef npy_cdouble __pyx_t_5numpy_cdouble_t;
  */
 typedef npy_clongdouble __pyx_t_5numpy_clongdouble_t;
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":773
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":773
  * ctypedef npy_clongdouble clongdouble_t
  * 
  * ctypedef npy_cdouble     complex_t             # <<<<<<<<<<<<<<
@@ -1673,7 +1673,7 @@ static PyObject *__pyx_codeobj__15;
 static PyObject *__pyx_codeobj__17;
 /* Late includes */
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":12
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":11
  * ctypedef np.float_t DTYPE_t
  * 
  * cdef p11_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1690,7 +1690,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p11_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":13
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":12
  * 
  * cdef p11_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = pow(sin(vza), 2) * pow(sin(iza), 2)             # <<<<<<<<<<<<<<
@@ -1699,7 +1699,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (pow(sin(__pyx_v_vza), 2.0) * pow(sin(__pyx_v_iza), 2.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":14
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":13
  * cdef p11_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = pow(sin(vza), 2) * pow(sin(iza), 2)
  *     cdef float second = pow(cos(vza), 2) * pow(cos(iza), 2) * pow(cos(raa), 2)             # <<<<<<<<<<<<<<
@@ -1708,7 +1708,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_second = ((pow(cos(__pyx_v_vza), 2.0) * pow(cos(__pyx_v_iza), 2.0)) * pow(cos(__pyx_v_raa), 2.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":15
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":14
  *     cdef float first = pow(sin(vza), 2) * pow(sin(iza), 2)
  *     cdef float second = pow(cos(vza), 2) * pow(cos(iza), 2) * pow(cos(raa), 2)
  *     cdef float third = 2 * sin(vza) * sin(iza) * cos(vza) * cos(iza) * cos(raa)             # <<<<<<<<<<<<<<
@@ -1717,7 +1717,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_third = (((((2.0 * sin(__pyx_v_vza)) * sin(__pyx_v_iza)) * cos(__pyx_v_vza)) * cos(__pyx_v_iza)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":17
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":16
  *     cdef float third = 2 * sin(vza) * sin(iza) * cos(vza) * cos(iza) * cos(raa)
  * 
  *     return first + second + third             # <<<<<<<<<<<<<<
@@ -1725,13 +1725,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p12_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(((__pyx_v_first + __pyx_v_second) + __pyx_v_third)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 17, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(((__pyx_v_first + __pyx_v_second) + __pyx_v_third)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 16, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":12
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":11
  * ctypedef np.float_t DTYPE_t
  * 
  * cdef p11_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1750,7 +1750,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":19
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":18
  *     return first + second + third
  * 
  * cdef p12_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1765,7 +1765,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p12_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":20
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":19
  * 
  * cdef p12_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = pow(cos(vza), 2) * pow(sin(raa), 2)             # <<<<<<<<<<<<<<
@@ -1774,7 +1774,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (pow(cos(__pyx_v_vza), 2.0) * pow(sin(__pyx_v_raa), 2.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":22
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":21
  *     cdef float first = pow(cos(vza), 2) * pow(sin(raa), 2)
  * 
  *     return first             # <<<<<<<<<<<<<<
@@ -1782,13 +1782,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p13_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 22, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 21, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":19
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":18
  *     return first + second + third
  * 
  * cdef p12_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1807,7 +1807,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":24
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":23
  *     return first
  * 
  * cdef p13_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1823,7 +1823,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p13_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":25
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":24
  * 
  * cdef p13_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = cos(vza) * sin(vza) * sin(iza) * sin(raa)             # <<<<<<<<<<<<<<
@@ -1832,7 +1832,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (((cos(__pyx_v_vza) * sin(__pyx_v_vza)) * sin(__pyx_v_iza)) * sin(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":26
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":25
  * cdef p13_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = cos(vza) * sin(vza) * sin(iza) * sin(raa)
  *     cdef float second = pow(cos(vza), 2) * cos(iza) * sin(raa) * cos(raa)             # <<<<<<<<<<<<<<
@@ -1841,7 +1841,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_second = (((pow(cos(__pyx_v_vza), 2.0) * cos(__pyx_v_iza)) * sin(__pyx_v_raa)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":28
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":27
  *     cdef float second = pow(cos(vza), 2) * cos(iza) * sin(raa) * cos(raa)
  * 
  *     return first + second             # <<<<<<<<<<<<<<
@@ -1849,13 +1849,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p21_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 28, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 27, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":24
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":23
  *     return first
  * 
  * cdef p13_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1874,7 +1874,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":30
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":29
  *     return first + second
  * 
  * cdef p21_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1889,7 +1889,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p21_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":31
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":30
  * 
  * cdef p21_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = pow(cos(iza), 2) * pow(sin(raa), 2)             # <<<<<<<<<<<<<<
@@ -1898,7 +1898,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (pow(cos(__pyx_v_iza), 2.0) * pow(sin(__pyx_v_raa), 2.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":33
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":32
  *     cdef float first = pow(cos(iza), 2) * pow(sin(raa), 2)
  * 
  *     return first             # <<<<<<<<<<<<<<
@@ -1906,13 +1906,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p22_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 33, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 32, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":30
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":29
  *     return first + second
  * 
  * cdef p21_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1931,7 +1931,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":35
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":34
  *     return first
  * 
  * cdef p22_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1946,7 +1946,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p22_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":36
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":35
  * 
  * cdef p22_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = pow(cos(raa), 2)             # <<<<<<<<<<<<<<
@@ -1955,7 +1955,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = pow(cos(__pyx_v_raa), 2.0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":38
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":37
  *     cdef float first = pow(cos(raa), 2)
  * 
  *     return first             # <<<<<<<<<<<<<<
@@ -1963,13 +1963,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p23_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 38, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 37, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":35
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":34
  *     return first
  * 
  * cdef p22_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -1988,7 +1988,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":40
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":39
  *     return first
  * 
  * cdef p23_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2003,7 +2003,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p23_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":41
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":40
  * 
  * cdef p23_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = -cos(iza) * sin(raa) * cos(raa)             # <<<<<<<<<<<<<<
@@ -2012,7 +2012,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (((-cos(__pyx_v_iza)) * sin(__pyx_v_raa)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":43
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":42
  *     cdef float first = -cos(iza) * sin(raa) * cos(raa)
  * 
  *     return first             # <<<<<<<<<<<<<<
@@ -2020,13 +2020,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p31_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 43, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 42, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":40
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":39
  *     return first
  * 
  * cdef p23_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2045,7 +2045,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":45
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":44
  *     return first
  * 
  * cdef p31_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2061,7 +2061,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p31_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":46
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":45
  * 
  * cdef p31_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = -2 * cos(vza) * sin(iza) * cos(iza) * sin(raa)             # <<<<<<<<<<<<<<
@@ -2070,7 +2070,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = ((((-2.0 * cos(__pyx_v_vza)) * sin(__pyx_v_iza)) * cos(__pyx_v_iza)) * sin(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":47
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":46
  * cdef p31_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = -2 * cos(vza) * sin(iza) * cos(iza) * sin(raa)
  *     cdef float second = - cos(vza) * pow(cos(iza), 2) * cos(raa) * sin(raa)             # <<<<<<<<<<<<<<
@@ -2079,7 +2079,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_second = ((((-cos(__pyx_v_vza)) * pow(cos(__pyx_v_iza), 2.0)) * cos(__pyx_v_raa)) * sin(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":49
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":48
  *     cdef float second = - cos(vza) * pow(cos(iza), 2) * cos(raa) * sin(raa)
  * 
  *     return first + second             # <<<<<<<<<<<<<<
@@ -2087,13 +2087,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p32_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 49, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 48, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":45
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":44
  *     return first
  * 
  * cdef p31_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2112,7 +2112,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":51
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":50
  *     return first + second
  * 
  * cdef p32_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2127,7 +2127,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p32_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":52
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":51
  * 
  * cdef p32_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = 2 * cos(vza) * sin(raa) * cos(raa)             # <<<<<<<<<<<<<<
@@ -2136,7 +2136,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = (((2.0 * cos(__pyx_v_vza)) * sin(__pyx_v_raa)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":54
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":53
  *     cdef float first = 2 * cos(vza) * sin(raa) * cos(raa)
  * 
  *     return first             # <<<<<<<<<<<<<<
@@ -2144,13 +2144,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p33_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 54, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble(__pyx_v_first); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 53, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":51
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":50
  *     return first + second
  * 
  * cdef p32_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2169,7 +2169,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":56
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":55
  *     return first
  * 
  * cdef p33_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2185,7 +2185,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p33_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":57
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":56
  * 
  * cdef p33_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = sin(vza) * sin(iza) * cos(raa)             # <<<<<<<<<<<<<<
@@ -2194,7 +2194,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = ((sin(__pyx_v_vza) * sin(__pyx_v_iza)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":58
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":57
  * cdef p33_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = sin(vza) * sin(iza) * cos(raa)
  *     cdef float second = cos(vza) * cos(iza) * cos(2 * raa)             # <<<<<<<<<<<<<<
@@ -2203,7 +2203,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_second = ((cos(__pyx_v_vza) * cos(__pyx_v_iza)) * cos((2.0 * __pyx_v_raa)));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":60
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":59
  *     cdef float second = cos(vza) * cos(iza) * cos(2 * raa)
  * 
  *     return first + second             # <<<<<<<<<<<<<<
@@ -2211,13 +2211,13 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  * cdef p44_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 60, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 59, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":56
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":55
  *     return first
  * 
  * cdef p33_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2236,7 +2236,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":62
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":61
  *     return first + second
  * 
  * cdef p44_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2252,7 +2252,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("p44_c", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":63
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":62
  * 
  * cdef p44_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = sin(vza) * sin(iza) * cos(raa)             # <<<<<<<<<<<<<<
@@ -2261,7 +2261,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_first = ((sin(__pyx_v_vza) * sin(__pyx_v_iza)) * cos(__pyx_v_raa));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":64
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":63
  * cdef p44_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):
  *     cdef float first = sin(vza) * sin(iza) * cos(raa)
  *     cdef float second = cos(vza) * cos(iza)             # <<<<<<<<<<<<<<
@@ -2270,21 +2270,21 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
  */
   __pyx_v_second = (cos(__pyx_v_vza) * cos(__pyx_v_iza));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":66
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":65
  *     cdef float second = cos(vza) * cos(iza)
  * 
  *     return first + second             # <<<<<<<<<<<<<<
  * 
- * 
+ * cdef pmatrix_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 66, __pyx_L1_error)
+  __pyx_t_1 = PyFloat_FromDouble((__pyx_v_first + __pyx_v_second)); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 65, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":62
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":61
  *     return first + second
  * 
  * cdef p44_c(DTYPE_t iza, DTYPE_t raa, DTYPE_t vza):             # <<<<<<<<<<<<<<
@@ -2303,8 +2303,8 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":69
- * 
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":67
+ *     return first + second
  * 
  * cdef pmatrix_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
@@ -2350,38 +2350,38 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_pybuffernd_mat.data = NULL;
   __pyx_pybuffernd_mat.rcbuffer = &__pyx_pybuffer_mat;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":70
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":68
  * 
  * cdef pmatrix_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)             # <<<<<<<<<<<<<<
  * 
  *     mat[0, 0] = p11_c(iza, raa, vza)
  */
-  __pyx_t_1 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_1 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  __pyx_t_2 = __Pyx_PyObject_GetAttrStr(__pyx_t_1, __pyx_n_s_zeros); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_2 = __Pyx_PyObject_GetAttrStr(__pyx_t_1, __pyx_n_s_zeros); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_2);
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
-  __pyx_t_1 = __Pyx_PyDict_NewPresized(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_1 = __Pyx_PyDict_NewPresized(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  __pyx_t_3 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_3)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_3 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_3)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_3);
-  __pyx_t_4 = __Pyx_PyObject_GetAttrStr(__pyx_t_3, __pyx_n_s_float); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_4 = __Pyx_PyObject_GetAttrStr(__pyx_t_3, __pyx_n_s_float); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
   __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
-  if (PyDict_SetItem(__pyx_t_1, __pyx_n_s_dtype, __pyx_t_4) < 0) __PYX_ERR(0, 70, __pyx_L1_error)
+  if (PyDict_SetItem(__pyx_t_1, __pyx_n_s_dtype, __pyx_t_4) < 0) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
-  __pyx_t_4 = __Pyx_PyObject_Call(__pyx_t_2, __pyx_tuple__2, __pyx_t_1); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_t_4 = __Pyx_PyObject_Call(__pyx_t_2, __pyx_tuple__2, __pyx_t_1); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
   __Pyx_DECREF(__pyx_t_2); __pyx_t_2 = 0;
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
-  if (!(likely(((__pyx_t_4) == Py_None) || likely(__Pyx_TypeTest(__pyx_t_4, __pyx_ptype_5numpy_ndarray))))) __PYX_ERR(0, 70, __pyx_L1_error)
+  if (!(likely(((__pyx_t_4) == Py_None) || likely(__Pyx_TypeTest(__pyx_t_4, __pyx_ptype_5numpy_ndarray))))) __PYX_ERR(0, 68, __pyx_L1_error)
   __pyx_t_5 = ((PyArrayObject *)__pyx_t_4);
   {
     __Pyx_BufFmt_StackElem __pyx_stack[1];
     if (unlikely(__Pyx_GetBufferAndValidate(&__pyx_pybuffernd_mat.rcbuffer->pybuffer, (PyObject*)__pyx_t_5, &__Pyx_TypeInfo_nn___pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t, PyBUF_FORMAT| PyBUF_STRIDES| PyBUF_WRITABLE, 2, 0, __pyx_stack) == -1)) {
       __pyx_v_mat = ((PyArrayObject *)Py_None); __Pyx_INCREF(Py_None); __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf = NULL;
-      __PYX_ERR(0, 70, __pyx_L1_error)
+      __PYX_ERR(0, 68, __pyx_L1_error)
     } else {__pyx_pybuffernd_mat.diminfo[0].strides = __pyx_pybuffernd_mat.rcbuffer->pybuffer.strides[0]; __pyx_pybuffernd_mat.diminfo[0].shape = __pyx_pybuffernd_mat.rcbuffer->pybuffer.shape[0]; __pyx_pybuffernd_mat.diminfo[1].strides = __pyx_pybuffernd_mat.rcbuffer->pybuffer.strides[1]; __pyx_pybuffernd_mat.diminfo[1].shape = __pyx_pybuffernd_mat.rcbuffer->pybuffer.shape[1];
     }
   }
@@ -2389,16 +2389,16 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_v_mat = ((PyArrayObject *)__pyx_t_4);
   __pyx_t_4 = 0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":72
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":70
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
  * 
  *     mat[0, 0] = p11_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[0, 1] = p12_c(iza, raa, vza)
  *     mat[0, 2] = p13_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p11_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 72, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p11_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 70, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 72, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 70, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_7 = 0;
   __pyx_t_8 = 0;
@@ -2413,20 +2413,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_8 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 72, __pyx_L1_error)
+    __PYX_ERR(0, 70, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_7, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_8, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":73
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":71
  * 
  *     mat[0, 0] = p11_c(iza, raa, vza)
  *     mat[0, 1] = p12_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[0, 2] = p13_c(iza, raa, vza)
  * 
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p12_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 73, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p12_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 71, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 73, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 71, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_10 = 0;
   __pyx_t_11 = 1;
@@ -2441,20 +2441,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_11 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 73, __pyx_L1_error)
+    __PYX_ERR(0, 71, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_10, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_11, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":74
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":72
  *     mat[0, 0] = p11_c(iza, raa, vza)
  *     mat[0, 1] = p12_c(iza, raa, vza)
  *     mat[0, 2] = p13_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  * 
  *     mat[1, 0] = p21_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p13_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 74, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p13_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 72, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 74, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 72, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_12 = 0;
   __pyx_t_13 = 2;
@@ -2469,20 +2469,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_13 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 74, __pyx_L1_error)
+    __PYX_ERR(0, 72, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_12, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_13, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":76
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":74
  *     mat[0, 2] = p13_c(iza, raa, vza)
  * 
  *     mat[1, 0] = p21_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[1, 1] = p22_c(iza, raa, vza)
  *     mat[1, 2] = p23_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p21_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 76, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p21_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 74, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 76, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 74, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_14 = 1;
   __pyx_t_15 = 0;
@@ -2497,20 +2497,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_15 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 76, __pyx_L1_error)
+    __PYX_ERR(0, 74, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_14, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_15, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":77
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":75
  * 
  *     mat[1, 0] = p21_c(iza, raa, vza)
  *     mat[1, 1] = p22_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[1, 2] = p23_c(iza, raa, vza)
  * 
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p22_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 77, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p22_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 75, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 77, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 75, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_16 = 1;
   __pyx_t_17 = 1;
@@ -2525,20 +2525,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_17 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 77, __pyx_L1_error)
+    __PYX_ERR(0, 75, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_16, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_17, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":78
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":76
  *     mat[1, 0] = p21_c(iza, raa, vza)
  *     mat[1, 1] = p22_c(iza, raa, vza)
  *     mat[1, 2] = p23_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  * 
  *     mat[2, 0] = p31_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p23_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 78, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p23_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 76, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 78, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 76, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_18 = 1;
   __pyx_t_19 = 2;
@@ -2553,20 +2553,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_19 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 78, __pyx_L1_error)
+    __PYX_ERR(0, 76, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_18, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_19, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":80
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":78
  *     mat[1, 2] = p23_c(iza, raa, vza)
  * 
  *     mat[2, 0] = p31_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[2, 1] = p32_c(iza, raa, vza)
  *     mat[2, 2] = p33_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p31_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 80, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p31_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 78, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 80, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 78, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_20 = 2;
   __pyx_t_21 = 0;
@@ -2581,20 +2581,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_21 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 80, __pyx_L1_error)
+    __PYX_ERR(0, 78, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_20, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_21, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":81
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":79
  * 
  *     mat[2, 0] = p31_c(iza, raa, vza)
  *     mat[2, 1] = p32_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  *     mat[2, 2] = p33_c(iza, raa, vza)
  * 
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p32_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 81, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p32_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 79, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 81, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 79, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_22 = 2;
   __pyx_t_23 = 1;
@@ -2609,20 +2609,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_23 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 81, __pyx_L1_error)
+    __PYX_ERR(0, 79, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_22, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_23, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":82
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":80
  *     mat[2, 0] = p31_c(iza, raa, vza)
  *     mat[2, 1] = p32_c(iza, raa, vza)
  *     mat[2, 2] = p33_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  * 
  *     mat[3, 3] = p44_c(iza, raa, vza)
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p33_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 82, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p33_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 80, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 82, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 80, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_24 = 2;
   __pyx_t_25 = 2;
@@ -2637,20 +2637,20 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_25 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 82, __pyx_L1_error)
+    __PYX_ERR(0, 80, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_24, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_25, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":84
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":82
  *     mat[2, 2] = p33_c(iza, raa, vza)
  * 
  *     mat[3, 3] = p44_c(iza, raa, vza)             # <<<<<<<<<<<<<<
  * 
  *     return mat
  */
-  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p44_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 84, __pyx_L1_error)
+  __pyx_t_4 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_p44_c(__pyx_v_iza, __pyx_v_raa, __pyx_v_vza); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 82, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
-  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 84, __pyx_L1_error)
+  __pyx_t_6 = __pyx_PyFloat_AsDouble(__pyx_t_4); if (unlikely((__pyx_t_6 == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 82, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
   __pyx_t_26 = 3;
   __pyx_t_27 = 3;
@@ -2665,11 +2665,11 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_27 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_9 = 1;
   if (unlikely(__pyx_t_9 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_9);
-    __PYX_ERR(0, 84, __pyx_L1_error)
+    __PYX_ERR(0, 82, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_26, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_27, __pyx_pybuffernd_mat.diminfo[1].strides) = __pyx_t_6;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":86
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":84
  *     mat[3, 3] = p44_c(iza, raa, vza)
  * 
  *     return mat             # <<<<<<<<<<<<<<
@@ -2681,8 +2681,8 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_r = ((PyObject *)__pyx_v_mat);
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":69
- * 
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":67
+ *     return first + second
  * 
  * cdef pmatrix_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
@@ -2713,12 +2713,12 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":88
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":86
  *     return mat
  * 
  * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- * 
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
+ * 
  */
 
 static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_dblquad_c(CYTHON_UNUSED __pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t __pyx_v_iza, __pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t __pyx_v_vza, CYTHON_UNUSED __pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t __pyx_v_raa) {
@@ -2747,38 +2747,38 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_pybuffernd_mat.data = NULL;
   __pyx_pybuffernd_mat.rcbuffer = &__pyx_pybuffer_mat;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":90
- * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":87
  * 
+ * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)             # <<<<<<<<<<<<<<
  * 
  *     mat[0, 0] = pi * sin(vza) ** 2 + pi / 3
  */
-  __pyx_t_1 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_1 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  __pyx_t_2 = __Pyx_PyObject_GetAttrStr(__pyx_t_1, __pyx_n_s_zeros); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_2 = __Pyx_PyObject_GetAttrStr(__pyx_t_1, __pyx_n_s_zeros); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_2);
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
-  __pyx_t_1 = __Pyx_PyDict_NewPresized(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_1 = __Pyx_PyDict_NewPresized(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  __pyx_t_3 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_3)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_3 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_3)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_3);
-  __pyx_t_4 = __Pyx_PyObject_GetAttrStr(__pyx_t_3, __pyx_n_s_float); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_4 = __Pyx_PyObject_GetAttrStr(__pyx_t_3, __pyx_n_s_float); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
   __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
-  if (PyDict_SetItem(__pyx_t_1, __pyx_n_s_dtype, __pyx_t_4) < 0) __PYX_ERR(0, 90, __pyx_L1_error)
+  if (PyDict_SetItem(__pyx_t_1, __pyx_n_s_dtype, __pyx_t_4) < 0) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
-  __pyx_t_4 = __Pyx_PyObject_Call(__pyx_t_2, __pyx_tuple__4, __pyx_t_1); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_t_4 = __Pyx_PyObject_Call(__pyx_t_2, __pyx_tuple__4, __pyx_t_1); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_4);
   __Pyx_DECREF(__pyx_t_2); __pyx_t_2 = 0;
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
-  if (!(likely(((__pyx_t_4) == Py_None) || likely(__Pyx_TypeTest(__pyx_t_4, __pyx_ptype_5numpy_ndarray))))) __PYX_ERR(0, 90, __pyx_L1_error)
+  if (!(likely(((__pyx_t_4) == Py_None) || likely(__Pyx_TypeTest(__pyx_t_4, __pyx_ptype_5numpy_ndarray))))) __PYX_ERR(0, 87, __pyx_L1_error)
   __pyx_t_5 = ((PyArrayObject *)__pyx_t_4);
   {
     __Pyx_BufFmt_StackElem __pyx_stack[1];
     if (unlikely(__Pyx_GetBufferAndValidate(&__pyx_pybuffernd_mat.rcbuffer->pybuffer, (PyObject*)__pyx_t_5, &__Pyx_TypeInfo_nn___pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t, PyBUF_FORMAT| PyBUF_STRIDES| PyBUF_WRITABLE, 2, 0, __pyx_stack) == -1)) {
       __pyx_v_mat = ((PyArrayObject *)Py_None); __Pyx_INCREF(Py_None); __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf = NULL;
-      __PYX_ERR(0, 90, __pyx_L1_error)
+      __PYX_ERR(0, 87, __pyx_L1_error)
     } else {__pyx_pybuffernd_mat.diminfo[0].strides = __pyx_pybuffernd_mat.rcbuffer->pybuffer.strides[0]; __pyx_pybuffernd_mat.diminfo[0].shape = __pyx_pybuffernd_mat.rcbuffer->pybuffer.shape[0]; __pyx_pybuffernd_mat.diminfo[1].strides = __pyx_pybuffernd_mat.rcbuffer->pybuffer.strides[1]; __pyx_pybuffernd_mat.diminfo[1].shape = __pyx_pybuffernd_mat.rcbuffer->pybuffer.shape[1];
     }
   }
@@ -2786,7 +2786,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_v_mat = ((PyArrayObject *)__pyx_t_4);
   __pyx_t_4 = 0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":92
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":89
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
  * 
  *     mat[0, 0] = pi * sin(vza) ** 2 + pi / 3             # <<<<<<<<<<<<<<
@@ -2806,11 +2806,11 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_7 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_8 = 1;
   if (unlikely(__pyx_t_8 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_8);
-    __PYX_ERR(0, 92, __pyx_L1_error)
+    __PYX_ERR(0, 89, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_6, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_7, __pyx_pybuffernd_mat.diminfo[1].strides) = ((M_PI * pow(sin(__pyx_v_vza), 2.0)) + (M_PI / 3.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":93
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":90
  * 
  *     mat[0, 0] = pi * sin(vza) ** 2 + pi / 3
  *     mat[0, 1] = pi * cos(vza) ** 2             # <<<<<<<<<<<<<<
@@ -2830,11 +2830,11 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_10 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_8 = 1;
   if (unlikely(__pyx_t_8 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_8);
-    __PYX_ERR(0, 93, __pyx_L1_error)
+    __PYX_ERR(0, 90, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_9, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_10, __pyx_pybuffernd_mat.diminfo[1].strides) = (M_PI * pow(cos(__pyx_v_vza), 2.0));
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":95
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":92
  *     mat[0, 1] = pi * cos(vza) ** 2
  * 
  *     mat[1, 0] = pi / 3             # <<<<<<<<<<<<<<
@@ -2854,11 +2854,11 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_12 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_8 = 1;
   if (unlikely(__pyx_t_8 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_8);
-    __PYX_ERR(0, 95, __pyx_L1_error)
+    __PYX_ERR(0, 92, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_11, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_12, __pyx_pybuffernd_mat.diminfo[1].strides) = (M_PI / 3.0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":96
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":93
  * 
  *     mat[1, 0] = pi / 3
  *     mat[1, 1] = pi             # <<<<<<<<<<<<<<
@@ -2878,11 +2878,11 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   } else if (unlikely(__pyx_t_14 >= __pyx_pybuffernd_mat.diminfo[1].shape)) __pyx_t_8 = 1;
   if (unlikely(__pyx_t_8 != -1)) {
     __Pyx_RaiseBufferIndexError(__pyx_t_8);
-    __PYX_ERR(0, 96, __pyx_L1_error)
+    __PYX_ERR(0, 93, __pyx_L1_error)
   }
   *__Pyx_BufPtrStrided2d(__pyx_t_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_DTYPE_t *, __pyx_pybuffernd_mat.rcbuffer->pybuffer.buf, __pyx_t_13, __pyx_pybuffernd_mat.diminfo[0].strides, __pyx_t_14, __pyx_pybuffernd_mat.diminfo[1].strides) = M_PI;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":98
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":95
  *     mat[1, 1] = pi
  * 
  *     return mat             # <<<<<<<<<<<<<<
@@ -2894,12 +2894,12 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   __pyx_r = ((PyObject *)__pyx_v_mat);
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":88
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":86
  *     return mat
  * 
  * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- * 
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)
+ * 
  */
 
   /* function exit code */
@@ -2926,7 +2926,7 @@ static PyObject *__pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":97
  *     return mat
  * 
  * def dblquad_c_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
@@ -2969,17 +2969,17 @@ static PyObject *__pyx_pw_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
         case  1:
         if (likely((values[1] = __Pyx_PyDict_GetItemStr(__pyx_kwds, __pyx_n_s_vza)) != 0)) kw_args--;
         else {
-          __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, 1); __PYX_ERR(0, 100, __pyx_L3_error)
+          __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, 1); __PYX_ERR(0, 97, __pyx_L3_error)
         }
         CYTHON_FALLTHROUGH;
         case  2:
         if (likely((values[2] = __Pyx_PyDict_GetItemStr(__pyx_kwds, __pyx_n_s_raa)) != 0)) kw_args--;
         else {
-          __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, 2); __PYX_ERR(0, 100, __pyx_L3_error)
+          __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, 2); __PYX_ERR(0, 97, __pyx_L3_error)
         }
       }
       if (unlikely(kw_args > 0)) {
-        if (unlikely(__Pyx_ParseOptionalKeywords(__pyx_kwds, __pyx_pyargnames, 0, values, pos_args, "dblquad_c_wrapper") < 0)) __PYX_ERR(0, 100, __pyx_L3_error)
+        if (unlikely(__Pyx_ParseOptionalKeywords(__pyx_kwds, __pyx_pyargnames, 0, values, pos_args, "dblquad_c_wrapper") < 0)) __PYX_ERR(0, 97, __pyx_L3_error)
       }
     } else if (PyTuple_GET_SIZE(__pyx_args) != 3) {
       goto __pyx_L5_argtuple_error;
@@ -2988,13 +2988,13 @@ static PyObject *__pyx_pw_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
       values[1] = PyTuple_GET_ITEM(__pyx_args, 1);
       values[2] = PyTuple_GET_ITEM(__pyx_args, 2);
     }
-    __pyx_v_iza = __pyx_PyFloat_AsDouble(values[0]); if (unlikely((__pyx_v_iza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
-    __pyx_v_vza = __pyx_PyFloat_AsDouble(values[1]); if (unlikely((__pyx_v_vza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
-    __pyx_v_raa = __pyx_PyFloat_AsDouble(values[2]); if (unlikely((__pyx_v_raa == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
+    __pyx_v_iza = __pyx_PyFloat_AsDouble(values[0]); if (unlikely((__pyx_v_iza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 97, __pyx_L3_error)
+    __pyx_v_vza = __pyx_PyFloat_AsDouble(values[1]); if (unlikely((__pyx_v_vza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 97, __pyx_L3_error)
+    __pyx_v_raa = __pyx_PyFloat_AsDouble(values[2]); if (unlikely((__pyx_v_raa == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 97, __pyx_L3_error)
   }
   goto __pyx_L4_argument_unpacking_done;
   __pyx_L5_argtuple_error:;
-  __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, PyTuple_GET_SIZE(__pyx_args)); __PYX_ERR(0, 100, __pyx_L3_error)
+  __Pyx_RaiseArgtupleInvalid("dblquad_c_wrapper", 1, 3, 3, PyTuple_GET_SIZE(__pyx_args)); __PYX_ERR(0, 97, __pyx_L3_error)
   __pyx_L3_error:;
   __Pyx_AddTraceback("pyrism.scattering.rayleigh.core.rayleigh_phase_c.dblquad_c_wrapper", __pyx_clineno, __pyx_lineno, __pyx_filename);
   __Pyx_RefNannyFinishContext();
@@ -3013,7 +3013,7 @@ static PyObject *__pyx_pf_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("dblquad_c_wrapper", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":101
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":98
  * 
  * def dblquad_c_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  *     return dblquad_c(iza, vza, raa)             # <<<<<<<<<<<<<<
@@ -3021,13 +3021,13 @@ static PyObject *__pyx_pf_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_dblquad_c(__pyx_v_iza, __pyx_v_vza, __pyx_v_raa); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 101, __pyx_L1_error)
+  __pyx_t_1 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_dblquad_c(__pyx_v_iza, __pyx_v_vza, __pyx_v_raa); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 98, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":97
  *     return mat
  * 
  * def dblquad_c_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
@@ -3046,11 +3046,11 @@ static PyObject *__pyx_pf_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
   return __pyx_r;
 }
 
-/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":103
+/* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
  *     return dblquad_c(iza, vza, raa)
  * 
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- *     return pmatrix_c( iza, vza, raa)
+ *     return pmatrix_c(iza, vza, raa)
  */
 
 /* Python wrapper */
@@ -3088,17 +3088,17 @@ static PyObject *__pyx_pw_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
         case  1:
         if (likely((values[1] = __Pyx_PyDict_GetItemStr(__pyx_kwds, __pyx_n_s_vza)) != 0)) kw_args--;
         else {
-          __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, 1); __PYX_ERR(0, 103, __pyx_L3_error)
+          __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, 1); __PYX_ERR(0, 100, __pyx_L3_error)
         }
         CYTHON_FALLTHROUGH;
         case  2:
         if (likely((values[2] = __Pyx_PyDict_GetItemStr(__pyx_kwds, __pyx_n_s_raa)) != 0)) kw_args--;
         else {
-          __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, 2); __PYX_ERR(0, 103, __pyx_L3_error)
+          __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, 2); __PYX_ERR(0, 100, __pyx_L3_error)
         }
       }
       if (unlikely(kw_args > 0)) {
-        if (unlikely(__Pyx_ParseOptionalKeywords(__pyx_kwds, __pyx_pyargnames, 0, values, pos_args, "pmatrix_wrapper") < 0)) __PYX_ERR(0, 103, __pyx_L3_error)
+        if (unlikely(__Pyx_ParseOptionalKeywords(__pyx_kwds, __pyx_pyargnames, 0, values, pos_args, "pmatrix_wrapper") < 0)) __PYX_ERR(0, 100, __pyx_L3_error)
       }
     } else if (PyTuple_GET_SIZE(__pyx_args) != 3) {
       goto __pyx_L5_argtuple_error;
@@ -3107,13 +3107,13 @@ static PyObject *__pyx_pw_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
       values[1] = PyTuple_GET_ITEM(__pyx_args, 1);
       values[2] = PyTuple_GET_ITEM(__pyx_args, 2);
     }
-    __pyx_v_iza = __pyx_PyFloat_AsDouble(values[0]); if (unlikely((__pyx_v_iza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 103, __pyx_L3_error)
-    __pyx_v_vza = __pyx_PyFloat_AsDouble(values[1]); if (unlikely((__pyx_v_vza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 103, __pyx_L3_error)
-    __pyx_v_raa = __pyx_PyFloat_AsDouble(values[2]); if (unlikely((__pyx_v_raa == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 103, __pyx_L3_error)
+    __pyx_v_iza = __pyx_PyFloat_AsDouble(values[0]); if (unlikely((__pyx_v_iza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
+    __pyx_v_vza = __pyx_PyFloat_AsDouble(values[1]); if (unlikely((__pyx_v_vza == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
+    __pyx_v_raa = __pyx_PyFloat_AsDouble(values[2]); if (unlikely((__pyx_v_raa == ((npy_double)-1)) && PyErr_Occurred())) __PYX_ERR(0, 100, __pyx_L3_error)
   }
   goto __pyx_L4_argument_unpacking_done;
   __pyx_L5_argtuple_error:;
-  __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, PyTuple_GET_SIZE(__pyx_args)); __PYX_ERR(0, 103, __pyx_L3_error)
+  __Pyx_RaiseArgtupleInvalid("pmatrix_wrapper", 1, 3, 3, PyTuple_GET_SIZE(__pyx_args)); __PYX_ERR(0, 100, __pyx_L3_error)
   __pyx_L3_error:;
   __Pyx_AddTraceback("pyrism.scattering.rayleigh.core.rayleigh_phase_c.pmatrix_wrapper", __pyx_clineno, __pyx_lineno, __pyx_filename);
   __Pyx_RefNannyFinishContext();
@@ -3132,23 +3132,23 @@ static PyObject *__pyx_pf_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("pmatrix_wrapper", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":104
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":101
  * 
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
- *     return pmatrix_c( iza, vza, raa)             # <<<<<<<<<<<<<<
+ *     return pmatrix_c(iza, vza, raa)             # <<<<<<<<<<<<<<
  */
   __Pyx_XDECREF(__pyx_r);
-  __pyx_t_1 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_pmatrix_c(__pyx_v_iza, __pyx_v_vza, __pyx_v_raa); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 104, __pyx_L1_error)
+  __pyx_t_1 = __pyx_f_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_pmatrix_c(__pyx_v_iza, __pyx_v_vza, __pyx_v_raa); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 101, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __pyx_r = __pyx_t_1;
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":103
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
  *     return dblquad_c(iza, vza, raa)
  * 
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- *     return pmatrix_c( iza, vza, raa)
+ *     return pmatrix_c(iza, vza, raa)
  */
 
   /* function exit code */
@@ -3162,7 +3162,7 @@ static PyObject *__pyx_pf_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":215
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":215
  *         # experimental exception made for __getbuffer__ and __releasebuffer__
  *         # -- the details of this may change.
  *         def __getbuffer__(ndarray self, Py_buffer* info, int flags):             # <<<<<<<<<<<<<<
@@ -3210,7 +3210,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_v_info->obj = Py_None; __Pyx_INCREF(Py_None);
   __Pyx_GIVEREF(__pyx_v_info->obj);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":222
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":222
  * 
  *             cdef int i, ndim
  *             cdef int endian_detector = 1             # <<<<<<<<<<<<<<
@@ -3219,7 +3219,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_endian_detector = 1;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":223
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":223
  *             cdef int i, ndim
  *             cdef int endian_detector = 1
  *             cdef bint little_endian = ((<char*>&endian_detector)[0] != 0)             # <<<<<<<<<<<<<<
@@ -3228,7 +3228,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_little_endian = ((((char *)(&__pyx_v_endian_detector))[0]) != 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":225
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":225
  *             cdef bint little_endian = ((<char*>&endian_detector)[0] != 0)
  * 
  *             ndim = PyArray_NDIM(self)             # <<<<<<<<<<<<<<
@@ -3237,7 +3237,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_ndim = PyArray_NDIM(__pyx_v_self);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
  *             ndim = PyArray_NDIM(self)
  * 
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3251,7 +3251,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     goto __pyx_L4_bool_binop_done;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":228
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":228
  * 
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_C_CONTIGUOUS)):             # <<<<<<<<<<<<<<
@@ -3262,7 +3262,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_t_1 = __pyx_t_2;
   __pyx_L4_bool_binop_done:;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
  *             ndim = PyArray_NDIM(self)
  * 
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3271,7 +3271,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   if (unlikely(__pyx_t_1)) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":229
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":229
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_C_CONTIGUOUS)):
  *                 raise ValueError(u"ndarray is not C contiguous")             # <<<<<<<<<<<<<<
@@ -3284,7 +3284,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
     __PYX_ERR(1, 229, __pyx_L1_error)
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":227
  *             ndim = PyArray_NDIM(self)
  * 
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3293,7 +3293,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
  *                 raise ValueError(u"ndarray is not C contiguous")
  * 
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3307,7 +3307,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     goto __pyx_L7_bool_binop_done;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":232
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":232
  * 
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_F_CONTIGUOUS)):             # <<<<<<<<<<<<<<
@@ -3318,7 +3318,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_t_1 = __pyx_t_2;
   __pyx_L7_bool_binop_done:;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
  *                 raise ValueError(u"ndarray is not C contiguous")
  * 
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3327,7 +3327,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   if (unlikely(__pyx_t_1)) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":233
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":233
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_F_CONTIGUOUS)):
  *                 raise ValueError(u"ndarray is not Fortran contiguous")             # <<<<<<<<<<<<<<
@@ -3340,7 +3340,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
     __PYX_ERR(1, 233, __pyx_L1_error)
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":231
  *                 raise ValueError(u"ndarray is not C contiguous")
  * 
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)             # <<<<<<<<<<<<<<
@@ -3349,7 +3349,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":235
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":235
  *                 raise ValueError(u"ndarray is not Fortran contiguous")
  * 
  *             info.buf = PyArray_DATA(self)             # <<<<<<<<<<<<<<
@@ -3358,7 +3358,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_info->buf = PyArray_DATA(__pyx_v_self);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":236
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":236
  * 
  *             info.buf = PyArray_DATA(self)
  *             info.ndim = ndim             # <<<<<<<<<<<<<<
@@ -3367,7 +3367,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_info->ndim = __pyx_v_ndim;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":237
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":237
  *             info.buf = PyArray_DATA(self)
  *             info.ndim = ndim
  *             if sizeof(npy_intp) != sizeof(Py_ssize_t):             # <<<<<<<<<<<<<<
@@ -3377,7 +3377,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_t_1 = (((sizeof(npy_intp)) != (sizeof(Py_ssize_t))) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":240
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":240
  *                 # Allocate new buffer for strides and shape info.
  *                 # This is allocated as one block, strides first.
  *                 info.strides = <Py_ssize_t*>PyObject_Malloc(sizeof(Py_ssize_t) * 2 * <size_t>ndim)             # <<<<<<<<<<<<<<
@@ -3386,7 +3386,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     __pyx_v_info->strides = ((Py_ssize_t *)PyObject_Malloc((((sizeof(Py_ssize_t)) * 2) * ((size_t)__pyx_v_ndim))));
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":241
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":241
  *                 # This is allocated as one block, strides first.
  *                 info.strides = <Py_ssize_t*>PyObject_Malloc(sizeof(Py_ssize_t) * 2 * <size_t>ndim)
  *                 info.shape = info.strides + ndim             # <<<<<<<<<<<<<<
@@ -3395,7 +3395,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     __pyx_v_info->shape = (__pyx_v_info->strides + __pyx_v_ndim);
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":242
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":242
  *                 info.strides = <Py_ssize_t*>PyObject_Malloc(sizeof(Py_ssize_t) * 2 * <size_t>ndim)
  *                 info.shape = info.strides + ndim
  *                 for i in range(ndim):             # <<<<<<<<<<<<<<
@@ -3407,7 +3407,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     for (__pyx_t_6 = 0; __pyx_t_6 < __pyx_t_5; __pyx_t_6+=1) {
       __pyx_v_i = __pyx_t_6;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":243
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":243
  *                 info.shape = info.strides + ndim
  *                 for i in range(ndim):
  *                     info.strides[i] = PyArray_STRIDES(self)[i]             # <<<<<<<<<<<<<<
@@ -3416,7 +3416,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
       (__pyx_v_info->strides[__pyx_v_i]) = (PyArray_STRIDES(__pyx_v_self)[__pyx_v_i]);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":244
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":244
  *                 for i in range(ndim):
  *                     info.strides[i] = PyArray_STRIDES(self)[i]
  *                     info.shape[i] = PyArray_DIMS(self)[i]             # <<<<<<<<<<<<<<
@@ -3426,7 +3426,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       (__pyx_v_info->shape[__pyx_v_i]) = (PyArray_DIMS(__pyx_v_self)[__pyx_v_i]);
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":237
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":237
  *             info.buf = PyArray_DATA(self)
  *             info.ndim = ndim
  *             if sizeof(npy_intp) != sizeof(Py_ssize_t):             # <<<<<<<<<<<<<<
@@ -3436,7 +3436,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     goto __pyx_L9;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":246
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":246
  *                     info.shape[i] = PyArray_DIMS(self)[i]
  *             else:
  *                 info.strides = <Py_ssize_t*>PyArray_STRIDES(self)             # <<<<<<<<<<<<<<
@@ -3446,7 +3446,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   /*else*/ {
     __pyx_v_info->strides = ((Py_ssize_t *)PyArray_STRIDES(__pyx_v_self));
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":247
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":247
  *             else:
  *                 info.strides = <Py_ssize_t*>PyArray_STRIDES(self)
  *                 info.shape = <Py_ssize_t*>PyArray_DIMS(self)             # <<<<<<<<<<<<<<
@@ -3457,7 +3457,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   }
   __pyx_L9:;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":248
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":248
  *                 info.strides = <Py_ssize_t*>PyArray_STRIDES(self)
  *                 info.shape = <Py_ssize_t*>PyArray_DIMS(self)
  *             info.suboffsets = NULL             # <<<<<<<<<<<<<<
@@ -3466,7 +3466,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_info->suboffsets = NULL;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":249
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":249
  *                 info.shape = <Py_ssize_t*>PyArray_DIMS(self)
  *             info.suboffsets = NULL
  *             info.itemsize = PyArray_ITEMSIZE(self)             # <<<<<<<<<<<<<<
@@ -3475,7 +3475,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_info->itemsize = PyArray_ITEMSIZE(__pyx_v_self);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":250
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":250
  *             info.suboffsets = NULL
  *             info.itemsize = PyArray_ITEMSIZE(self)
  *             info.readonly = not PyArray_ISWRITEABLE(self)             # <<<<<<<<<<<<<<
@@ -3484,7 +3484,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_info->readonly = (!(PyArray_ISWRITEABLE(__pyx_v_self) != 0));
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":253
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":253
  * 
  *             cdef int t
  *             cdef char* f = NULL             # <<<<<<<<<<<<<<
@@ -3493,7 +3493,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   __pyx_v_f = NULL;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":254
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":254
  *             cdef int t
  *             cdef char* f = NULL
  *             cdef dtype descr = self.descr             # <<<<<<<<<<<<<<
@@ -3505,7 +3505,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_v_descr = ((PyArray_Descr *)__pyx_t_3);
   __pyx_t_3 = 0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":257
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":257
  *             cdef int offset
  * 
  *             info.obj = self             # <<<<<<<<<<<<<<
@@ -3518,7 +3518,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __Pyx_DECREF(__pyx_v_info->obj);
   __pyx_v_info->obj = ((PyObject *)__pyx_v_self);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":259
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":259
  *             info.obj = self
  * 
  *             if not PyDataType_HASFIELDS(descr):             # <<<<<<<<<<<<<<
@@ -3528,7 +3528,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   __pyx_t_1 = ((!(PyDataType_HASFIELDS(__pyx_v_descr) != 0)) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":260
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":260
  * 
  *             if not PyDataType_HASFIELDS(descr):
  *                 t = descr.type_num             # <<<<<<<<<<<<<<
@@ -3538,7 +3538,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __pyx_t_4 = __pyx_v_descr->type_num;
     __pyx_v_t = __pyx_t_4;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
  *             if not PyDataType_HASFIELDS(descr):
  *                 t = descr.type_num
  *                 if ((descr.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -3558,7 +3558,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     }
     __pyx_L15_next_or:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":262
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":262
  *                 t = descr.type_num
  *                 if ((descr.byteorder == c'>' and little_endian) or
  *                     (descr.byteorder == c'<' and not little_endian)):             # <<<<<<<<<<<<<<
@@ -3575,7 +3575,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __pyx_t_1 = __pyx_t_2;
     __pyx_L14_bool_binop_done:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
  *             if not PyDataType_HASFIELDS(descr):
  *                 t = descr.type_num
  *                 if ((descr.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -3584,7 +3584,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     if (unlikely(__pyx_t_1)) {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":263
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":263
  *                 if ((descr.byteorder == c'>' and little_endian) or
  *                     (descr.byteorder == c'<' and not little_endian)):
  *                     raise ValueError(u"Non-native byte order not supported")             # <<<<<<<<<<<<<<
@@ -3597,7 +3597,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
       __PYX_ERR(1, 263, __pyx_L1_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":261
  *             if not PyDataType_HASFIELDS(descr):
  *                 t = descr.type_num
  *                 if ((descr.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -3606,7 +3606,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":264
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":264
  *                     (descr.byteorder == c'<' and not little_endian)):
  *                     raise ValueError(u"Non-native byte order not supported")
  *                 if   t == NPY_BYTE:        f = "b"             # <<<<<<<<<<<<<<
@@ -3618,7 +3618,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"b");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":265
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":265
  *                     raise ValueError(u"Non-native byte order not supported")
  *                 if   t == NPY_BYTE:        f = "b"
  *                 elif t == NPY_UBYTE:       f = "B"             # <<<<<<<<<<<<<<
@@ -3629,7 +3629,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"B");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":266
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":266
  *                 if   t == NPY_BYTE:        f = "b"
  *                 elif t == NPY_UBYTE:       f = "B"
  *                 elif t == NPY_SHORT:       f = "h"             # <<<<<<<<<<<<<<
@@ -3640,7 +3640,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"h");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":267
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":267
  *                 elif t == NPY_UBYTE:       f = "B"
  *                 elif t == NPY_SHORT:       f = "h"
  *                 elif t == NPY_USHORT:      f = "H"             # <<<<<<<<<<<<<<
@@ -3651,7 +3651,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"H");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":268
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":268
  *                 elif t == NPY_SHORT:       f = "h"
  *                 elif t == NPY_USHORT:      f = "H"
  *                 elif t == NPY_INT:         f = "i"             # <<<<<<<<<<<<<<
@@ -3662,7 +3662,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"i");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":269
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":269
  *                 elif t == NPY_USHORT:      f = "H"
  *                 elif t == NPY_INT:         f = "i"
  *                 elif t == NPY_UINT:        f = "I"             # <<<<<<<<<<<<<<
@@ -3673,7 +3673,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"I");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":270
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":270
  *                 elif t == NPY_INT:         f = "i"
  *                 elif t == NPY_UINT:        f = "I"
  *                 elif t == NPY_LONG:        f = "l"             # <<<<<<<<<<<<<<
@@ -3684,7 +3684,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"l");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":271
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":271
  *                 elif t == NPY_UINT:        f = "I"
  *                 elif t == NPY_LONG:        f = "l"
  *                 elif t == NPY_ULONG:       f = "L"             # <<<<<<<<<<<<<<
@@ -3695,7 +3695,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"L");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":272
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":272
  *                 elif t == NPY_LONG:        f = "l"
  *                 elif t == NPY_ULONG:       f = "L"
  *                 elif t == NPY_LONGLONG:    f = "q"             # <<<<<<<<<<<<<<
@@ -3706,7 +3706,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"q");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":273
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":273
  *                 elif t == NPY_ULONG:       f = "L"
  *                 elif t == NPY_LONGLONG:    f = "q"
  *                 elif t == NPY_ULONGLONG:   f = "Q"             # <<<<<<<<<<<<<<
@@ -3717,7 +3717,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"Q");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":274
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":274
  *                 elif t == NPY_LONGLONG:    f = "q"
  *                 elif t == NPY_ULONGLONG:   f = "Q"
  *                 elif t == NPY_FLOAT:       f = "f"             # <<<<<<<<<<<<<<
@@ -3728,7 +3728,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"f");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":275
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":275
  *                 elif t == NPY_ULONGLONG:   f = "Q"
  *                 elif t == NPY_FLOAT:       f = "f"
  *                 elif t == NPY_DOUBLE:      f = "d"             # <<<<<<<<<<<<<<
@@ -3739,7 +3739,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"d");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":276
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":276
  *                 elif t == NPY_FLOAT:       f = "f"
  *                 elif t == NPY_DOUBLE:      f = "d"
  *                 elif t == NPY_LONGDOUBLE:  f = "g"             # <<<<<<<<<<<<<<
@@ -3750,7 +3750,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"g");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":277
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":277
  *                 elif t == NPY_DOUBLE:      f = "d"
  *                 elif t == NPY_LONGDOUBLE:  f = "g"
  *                 elif t == NPY_CFLOAT:      f = "Zf"             # <<<<<<<<<<<<<<
@@ -3761,7 +3761,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"Zf");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":278
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":278
  *                 elif t == NPY_LONGDOUBLE:  f = "g"
  *                 elif t == NPY_CFLOAT:      f = "Zf"
  *                 elif t == NPY_CDOUBLE:     f = "Zd"             # <<<<<<<<<<<<<<
@@ -3772,7 +3772,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"Zd");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":279
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":279
  *                 elif t == NPY_CFLOAT:      f = "Zf"
  *                 elif t == NPY_CDOUBLE:     f = "Zd"
  *                 elif t == NPY_CLONGDOUBLE: f = "Zg"             # <<<<<<<<<<<<<<
@@ -3783,7 +3783,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       __pyx_v_f = ((char *)"Zg");
       break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":280
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":280
  *                 elif t == NPY_CDOUBLE:     f = "Zd"
  *                 elif t == NPY_CLONGDOUBLE: f = "Zg"
  *                 elif t == NPY_OBJECT:      f = "O"             # <<<<<<<<<<<<<<
@@ -3795,7 +3795,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       break;
       default:
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":282
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":282
  *                 elif t == NPY_OBJECT:      f = "O"
  *                 else:
  *                     raise ValueError(u"unknown dtype code in numpy.pxd (%d)" % t)             # <<<<<<<<<<<<<<
@@ -3816,7 +3816,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
       break;
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":283
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":283
  *                 else:
  *                     raise ValueError(u"unknown dtype code in numpy.pxd (%d)" % t)
  *                 info.format = f             # <<<<<<<<<<<<<<
@@ -3825,7 +3825,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     __pyx_v_info->format = __pyx_v_f;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":284
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":284
  *                     raise ValueError(u"unknown dtype code in numpy.pxd (%d)" % t)
  *                 info.format = f
  *                 return             # <<<<<<<<<<<<<<
@@ -3835,7 +3835,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __pyx_r = 0;
     goto __pyx_L0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":259
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":259
  *             info.obj = self
  * 
  *             if not PyDataType_HASFIELDS(descr):             # <<<<<<<<<<<<<<
@@ -3844,7 +3844,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":286
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":286
  *                 return
  *             else:
  *                 info.format = <char*>PyObject_Malloc(_buffer_format_string_len)             # <<<<<<<<<<<<<<
@@ -3854,7 +3854,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   /*else*/ {
     __pyx_v_info->format = ((char *)PyObject_Malloc(0xFF));
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":287
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":287
  *             else:
  *                 info.format = <char*>PyObject_Malloc(_buffer_format_string_len)
  *                 info.format[0] = c'^' # Native data types, manual alignment             # <<<<<<<<<<<<<<
@@ -3863,7 +3863,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     (__pyx_v_info->format[0]) = '^';
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":288
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":288
  *                 info.format = <char*>PyObject_Malloc(_buffer_format_string_len)
  *                 info.format[0] = c'^' # Native data types, manual alignment
  *                 offset = 0             # <<<<<<<<<<<<<<
@@ -3872,7 +3872,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
  */
     __pyx_v_offset = 0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":289
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":289
  *                 info.format[0] = c'^' # Native data types, manual alignment
  *                 offset = 0
  *                 f = _util_dtypestring(descr, info.format + 1,             # <<<<<<<<<<<<<<
@@ -3882,7 +3882,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     __pyx_t_8 = __pyx_f_5numpy__util_dtypestring(__pyx_v_descr, (__pyx_v_info->format + 1), (__pyx_v_info->format + 0xFF), (&__pyx_v_offset)); if (unlikely(__pyx_t_8 == ((char *)NULL))) __PYX_ERR(1, 289, __pyx_L1_error)
     __pyx_v_f = __pyx_t_8;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":292
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":292
  *                                       info.format + _buffer_format_string_len,
  *                                       &offset)
  *                 f[0] = c'\0' # Terminate format string             # <<<<<<<<<<<<<<
@@ -3892,7 +3892,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
     (__pyx_v_f[0]) = '\x00';
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":215
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":215
  *         # experimental exception made for __getbuffer__ and __releasebuffer__
  *         # -- the details of this may change.
  *         def __getbuffer__(ndarray self, Py_buffer* info, int flags):             # <<<<<<<<<<<<<<
@@ -3924,7 +3924,7 @@ static int __pyx_pf_5numpy_7ndarray___getbuffer__(PyArrayObject *__pyx_v_self, P
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":294
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":294
  *                 f[0] = c'\0' # Terminate format string
  * 
  *         def __releasebuffer__(ndarray self, Py_buffer* info):             # <<<<<<<<<<<<<<
@@ -3948,7 +3948,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
   int __pyx_t_1;
   __Pyx_RefNannySetupContext("__releasebuffer__", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":295
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":295
  * 
  *         def __releasebuffer__(ndarray self, Py_buffer* info):
  *             if PyArray_HASFIELDS(self):             # <<<<<<<<<<<<<<
@@ -3958,7 +3958,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
   __pyx_t_1 = (PyArray_HASFIELDS(__pyx_v_self) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":296
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":296
  *         def __releasebuffer__(ndarray self, Py_buffer* info):
  *             if PyArray_HASFIELDS(self):
  *                 PyObject_Free(info.format)             # <<<<<<<<<<<<<<
@@ -3967,7 +3967,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
  */
     PyObject_Free(__pyx_v_info->format);
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":295
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":295
  * 
  *         def __releasebuffer__(ndarray self, Py_buffer* info):
  *             if PyArray_HASFIELDS(self):             # <<<<<<<<<<<<<<
@@ -3976,7 +3976,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":297
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":297
  *             if PyArray_HASFIELDS(self):
  *                 PyObject_Free(info.format)
  *             if sizeof(npy_intp) != sizeof(Py_ssize_t):             # <<<<<<<<<<<<<<
@@ -3986,7 +3986,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
   __pyx_t_1 = (((sizeof(npy_intp)) != (sizeof(Py_ssize_t))) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":298
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":298
  *                 PyObject_Free(info.format)
  *             if sizeof(npy_intp) != sizeof(Py_ssize_t):
  *                 PyObject_Free(info.strides)             # <<<<<<<<<<<<<<
@@ -3995,7 +3995,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
  */
     PyObject_Free(__pyx_v_info->strides);
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":297
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":297
  *             if PyArray_HASFIELDS(self):
  *                 PyObject_Free(info.format)
  *             if sizeof(npy_intp) != sizeof(Py_ssize_t):             # <<<<<<<<<<<<<<
@@ -4004,7 +4004,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":294
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":294
  *                 f[0] = c'\0' # Terminate format string
  * 
  *         def __releasebuffer__(ndarray self, Py_buffer* info):             # <<<<<<<<<<<<<<
@@ -4016,7 +4016,7 @@ static void __pyx_pf_5numpy_7ndarray_2__releasebuffer__(PyArrayObject *__pyx_v_s
   __Pyx_RefNannyFinishContext();
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":775
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":775
  * ctypedef npy_cdouble     complex_t
  * 
  * cdef inline object PyArray_MultiIterNew1(a):             # <<<<<<<<<<<<<<
@@ -4030,7 +4030,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew1(PyObject *__
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("PyArray_MultiIterNew1", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":776
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":776
  * 
  * cdef inline object PyArray_MultiIterNew1(a):
  *     return PyArray_MultiIterNew(1, <void*>a)             # <<<<<<<<<<<<<<
@@ -4044,7 +4044,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew1(PyObject *__
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":775
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":775
  * ctypedef npy_cdouble     complex_t
  * 
  * cdef inline object PyArray_MultiIterNew1(a):             # <<<<<<<<<<<<<<
@@ -4063,7 +4063,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew1(PyObject *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":778
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":778
  *     return PyArray_MultiIterNew(1, <void*>a)
  * 
  * cdef inline object PyArray_MultiIterNew2(a, b):             # <<<<<<<<<<<<<<
@@ -4077,7 +4077,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew2(PyObject *__
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("PyArray_MultiIterNew2", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":779
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":779
  * 
  * cdef inline object PyArray_MultiIterNew2(a, b):
  *     return PyArray_MultiIterNew(2, <void*>a, <void*>b)             # <<<<<<<<<<<<<<
@@ -4091,7 +4091,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew2(PyObject *__
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":778
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":778
  *     return PyArray_MultiIterNew(1, <void*>a)
  * 
  * cdef inline object PyArray_MultiIterNew2(a, b):             # <<<<<<<<<<<<<<
@@ -4110,7 +4110,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew2(PyObject *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":781
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":781
  *     return PyArray_MultiIterNew(2, <void*>a, <void*>b)
  * 
  * cdef inline object PyArray_MultiIterNew3(a, b, c):             # <<<<<<<<<<<<<<
@@ -4124,7 +4124,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew3(PyObject *__
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("PyArray_MultiIterNew3", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":782
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":782
  * 
  * cdef inline object PyArray_MultiIterNew3(a, b, c):
  *     return PyArray_MultiIterNew(3, <void*>a, <void*>b, <void*> c)             # <<<<<<<<<<<<<<
@@ -4138,7 +4138,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew3(PyObject *__
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":781
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":781
  *     return PyArray_MultiIterNew(2, <void*>a, <void*>b)
  * 
  * cdef inline object PyArray_MultiIterNew3(a, b, c):             # <<<<<<<<<<<<<<
@@ -4157,7 +4157,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew3(PyObject *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":784
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":784
  *     return PyArray_MultiIterNew(3, <void*>a, <void*>b, <void*> c)
  * 
  * cdef inline object PyArray_MultiIterNew4(a, b, c, d):             # <<<<<<<<<<<<<<
@@ -4171,7 +4171,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew4(PyObject *__
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("PyArray_MultiIterNew4", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":785
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":785
  * 
  * cdef inline object PyArray_MultiIterNew4(a, b, c, d):
  *     return PyArray_MultiIterNew(4, <void*>a, <void*>b, <void*>c, <void*> d)             # <<<<<<<<<<<<<<
@@ -4185,7 +4185,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew4(PyObject *__
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":784
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":784
  *     return PyArray_MultiIterNew(3, <void*>a, <void*>b, <void*> c)
  * 
  * cdef inline object PyArray_MultiIterNew4(a, b, c, d):             # <<<<<<<<<<<<<<
@@ -4204,7 +4204,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew4(PyObject *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":787
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":787
  *     return PyArray_MultiIterNew(4, <void*>a, <void*>b, <void*>c, <void*> d)
  * 
  * cdef inline object PyArray_MultiIterNew5(a, b, c, d, e):             # <<<<<<<<<<<<<<
@@ -4218,7 +4218,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew5(PyObject *__
   PyObject *__pyx_t_1 = NULL;
   __Pyx_RefNannySetupContext("PyArray_MultiIterNew5", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":788
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":788
  * 
  * cdef inline object PyArray_MultiIterNew5(a, b, c, d, e):
  *     return PyArray_MultiIterNew(5, <void*>a, <void*>b, <void*>c, <void*> d, <void*> e)             # <<<<<<<<<<<<<<
@@ -4232,7 +4232,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew5(PyObject *__
   __pyx_t_1 = 0;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":787
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":787
  *     return PyArray_MultiIterNew(4, <void*>a, <void*>b, <void*>c, <void*> d)
  * 
  * cdef inline object PyArray_MultiIterNew5(a, b, c, d, e):             # <<<<<<<<<<<<<<
@@ -4251,7 +4251,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyArray_MultiIterNew5(PyObject *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":790
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":790
  *     return PyArray_MultiIterNew(5, <void*>a, <void*>b, <void*>c, <void*> d, <void*> e)
  * 
  * cdef inline tuple PyDataType_SHAPE(dtype d):             # <<<<<<<<<<<<<<
@@ -4265,7 +4265,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
   int __pyx_t_1;
   __Pyx_RefNannySetupContext("PyDataType_SHAPE", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":791
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":791
  * 
  * cdef inline tuple PyDataType_SHAPE(dtype d):
  *     if PyDataType_HASSUBARRAY(d):             # <<<<<<<<<<<<<<
@@ -4275,7 +4275,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
   __pyx_t_1 = (PyDataType_HASSUBARRAY(__pyx_v_d) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":792
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":792
  * cdef inline tuple PyDataType_SHAPE(dtype d):
  *     if PyDataType_HASSUBARRAY(d):
  *         return <tuple>d.subarray.shape             # <<<<<<<<<<<<<<
@@ -4287,7 +4287,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
     __pyx_r = ((PyObject*)__pyx_v_d->subarray->shape);
     goto __pyx_L0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":791
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":791
  * 
  * cdef inline tuple PyDataType_SHAPE(dtype d):
  *     if PyDataType_HASSUBARRAY(d):             # <<<<<<<<<<<<<<
@@ -4296,7 +4296,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":794
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":794
  *         return <tuple>d.subarray.shape
  *     else:
  *         return ()             # <<<<<<<<<<<<<<
@@ -4310,7 +4310,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
     goto __pyx_L0;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":790
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":790
  *     return PyArray_MultiIterNew(5, <void*>a, <void*>b, <void*>c, <void*> d, <void*> e)
  * 
  * cdef inline tuple PyDataType_SHAPE(dtype d):             # <<<<<<<<<<<<<<
@@ -4325,7 +4325,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_PyDataType_SHAPE(PyArray_Descr *__
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":796
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":796
  *         return ()
  * 
  * cdef inline char* _util_dtypestring(dtype descr, char* f, char* end, int* offset) except NULL:             # <<<<<<<<<<<<<<
@@ -4354,7 +4354,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
   char *__pyx_t_9;
   __Pyx_RefNannySetupContext("_util_dtypestring", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":801
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":801
  * 
  *     cdef dtype child
  *     cdef int endian_detector = 1             # <<<<<<<<<<<<<<
@@ -4363,7 +4363,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
   __pyx_v_endian_detector = 1;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":802
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":802
  *     cdef dtype child
  *     cdef int endian_detector = 1
  *     cdef bint little_endian = ((<char*>&endian_detector)[0] != 0)             # <<<<<<<<<<<<<<
@@ -4372,7 +4372,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
   __pyx_v_little_endian = ((((char *)(&__pyx_v_endian_detector))[0]) != 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":805
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":805
  *     cdef tuple fields
  * 
  *     for childname in descr.names:             # <<<<<<<<<<<<<<
@@ -4395,7 +4395,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __Pyx_XDECREF_SET(__pyx_v_childname, __pyx_t_3);
     __pyx_t_3 = 0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":806
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":806
  * 
  *     for childname in descr.names:
  *         fields = descr.fields[childname]             # <<<<<<<<<<<<<<
@@ -4412,7 +4412,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __Pyx_XDECREF_SET(__pyx_v_fields, ((PyObject*)__pyx_t_3));
     __pyx_t_3 = 0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":807
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":807
  *     for childname in descr.names:
  *         fields = descr.fields[childname]
  *         child, new_offset = fields             # <<<<<<<<<<<<<<
@@ -4447,7 +4447,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __Pyx_XDECREF_SET(__pyx_v_new_offset, __pyx_t_4);
     __pyx_t_4 = 0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":809
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":809
  *         child, new_offset = fields
  * 
  *         if (end - f) - <int>(new_offset - offset[0]) < 15:             # <<<<<<<<<<<<<<
@@ -4464,7 +4464,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __pyx_t_6 = ((((__pyx_v_end - __pyx_v_f) - ((int)__pyx_t_5)) < 15) != 0);
     if (unlikely(__pyx_t_6)) {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":810
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":810
  * 
  *         if (end - f) - <int>(new_offset - offset[0]) < 15:
  *             raise RuntimeError(u"Format string allocated too short, see comment in numpy.pxd")             # <<<<<<<<<<<<<<
@@ -4477,7 +4477,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
       __PYX_ERR(1, 810, __pyx_L1_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":809
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":809
  *         child, new_offset = fields
  * 
  *         if (end - f) - <int>(new_offset - offset[0]) < 15:             # <<<<<<<<<<<<<<
@@ -4486,7 +4486,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
  *             raise RuntimeError(u"Format string allocated too short, see comment in numpy.pxd")
  * 
  *         if ((child.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -4506,7 +4506,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     }
     __pyx_L8_next_or:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":813
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":813
  * 
  *         if ((child.byteorder == c'>' and little_endian) or
  *             (child.byteorder == c'<' and not little_endian)):             # <<<<<<<<<<<<<<
@@ -4523,7 +4523,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __pyx_t_6 = __pyx_t_7;
     __pyx_L7_bool_binop_done:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
  *             raise RuntimeError(u"Format string allocated too short, see comment in numpy.pxd")
  * 
  *         if ((child.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -4532,7 +4532,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
     if (unlikely(__pyx_t_6)) {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":814
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":814
  *         if ((child.byteorder == c'>' and little_endian) or
  *             (child.byteorder == c'<' and not little_endian)):
  *             raise ValueError(u"Non-native byte order not supported")             # <<<<<<<<<<<<<<
@@ -4545,7 +4545,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       __Pyx_DECREF(__pyx_t_3); __pyx_t_3 = 0;
       __PYX_ERR(1, 814, __pyx_L1_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":812
  *             raise RuntimeError(u"Format string allocated too short, see comment in numpy.pxd")
  * 
  *         if ((child.byteorder == c'>' and little_endian) or             # <<<<<<<<<<<<<<
@@ -4554,7 +4554,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":824
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":824
  * 
  *         # Output padding bytes
  *         while offset[0] < new_offset:             # <<<<<<<<<<<<<<
@@ -4570,7 +4570,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
       if (!__pyx_t_6) break;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":825
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":825
  *         # Output padding bytes
  *         while offset[0] < new_offset:
  *             f[0] = 120 # "x"; pad byte             # <<<<<<<<<<<<<<
@@ -4579,7 +4579,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
       (__pyx_v_f[0]) = 0x78;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":826
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":826
  *         while offset[0] < new_offset:
  *             f[0] = 120 # "x"; pad byte
  *             f += 1             # <<<<<<<<<<<<<<
@@ -4588,7 +4588,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
       __pyx_v_f = (__pyx_v_f + 1);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":827
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":827
  *             f[0] = 120 # "x"; pad byte
  *             f += 1
  *             offset[0] += 1             # <<<<<<<<<<<<<<
@@ -4599,7 +4599,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       (__pyx_v_offset[__pyx_t_8]) = ((__pyx_v_offset[__pyx_t_8]) + 1);
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":829
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":829
  *             offset[0] += 1
  * 
  *         offset[0] += child.itemsize             # <<<<<<<<<<<<<<
@@ -4609,7 +4609,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __pyx_t_8 = 0;
     (__pyx_v_offset[__pyx_t_8]) = ((__pyx_v_offset[__pyx_t_8]) + __pyx_v_child->elsize);
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":831
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":831
  *         offset[0] += child.itemsize
  * 
  *         if not PyDataType_HASFIELDS(child):             # <<<<<<<<<<<<<<
@@ -4619,7 +4619,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     __pyx_t_6 = ((!(PyDataType_HASFIELDS(__pyx_v_child) != 0)) != 0);
     if (__pyx_t_6) {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":832
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":832
  * 
  *         if not PyDataType_HASFIELDS(child):
  *             t = child.type_num             # <<<<<<<<<<<<<<
@@ -4631,7 +4631,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       __Pyx_XDECREF_SET(__pyx_v_t, __pyx_t_4);
       __pyx_t_4 = 0;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":833
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":833
  *         if not PyDataType_HASFIELDS(child):
  *             t = child.type_num
  *             if end - f < 5:             # <<<<<<<<<<<<<<
@@ -4641,7 +4641,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       __pyx_t_6 = (((__pyx_v_end - __pyx_v_f) < 5) != 0);
       if (unlikely(__pyx_t_6)) {
 
-        /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":834
+        /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":834
  *             t = child.type_num
  *             if end - f < 5:
  *                 raise RuntimeError(u"Format string allocated too short.")             # <<<<<<<<<<<<<<
@@ -4654,7 +4654,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         __Pyx_DECREF(__pyx_t_4); __pyx_t_4 = 0;
         __PYX_ERR(1, 834, __pyx_L1_error)
 
-        /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":833
+        /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":833
  *         if not PyDataType_HASFIELDS(child):
  *             t = child.type_num
  *             if end - f < 5:             # <<<<<<<<<<<<<<
@@ -4663,7 +4663,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":837
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":837
  * 
  *             # Until ticket #99 is fixed, use integers to avoid warnings
  *             if   t == NPY_BYTE:        f[0] =  98 #"b"             # <<<<<<<<<<<<<<
@@ -4681,7 +4681,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":838
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":838
  *             # Until ticket #99 is fixed, use integers to avoid warnings
  *             if   t == NPY_BYTE:        f[0] =  98 #"b"
  *             elif t == NPY_UBYTE:       f[0] =  66 #"B"             # <<<<<<<<<<<<<<
@@ -4699,7 +4699,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":839
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":839
  *             if   t == NPY_BYTE:        f[0] =  98 #"b"
  *             elif t == NPY_UBYTE:       f[0] =  66 #"B"
  *             elif t == NPY_SHORT:       f[0] = 104 #"h"             # <<<<<<<<<<<<<<
@@ -4717,7 +4717,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":840
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":840
  *             elif t == NPY_UBYTE:       f[0] =  66 #"B"
  *             elif t == NPY_SHORT:       f[0] = 104 #"h"
  *             elif t == NPY_USHORT:      f[0] =  72 #"H"             # <<<<<<<<<<<<<<
@@ -4735,7 +4735,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":841
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":841
  *             elif t == NPY_SHORT:       f[0] = 104 #"h"
  *             elif t == NPY_USHORT:      f[0] =  72 #"H"
  *             elif t == NPY_INT:         f[0] = 105 #"i"             # <<<<<<<<<<<<<<
@@ -4753,7 +4753,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":842
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":842
  *             elif t == NPY_USHORT:      f[0] =  72 #"H"
  *             elif t == NPY_INT:         f[0] = 105 #"i"
  *             elif t == NPY_UINT:        f[0] =  73 #"I"             # <<<<<<<<<<<<<<
@@ -4771,7 +4771,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":843
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":843
  *             elif t == NPY_INT:         f[0] = 105 #"i"
  *             elif t == NPY_UINT:        f[0] =  73 #"I"
  *             elif t == NPY_LONG:        f[0] = 108 #"l"             # <<<<<<<<<<<<<<
@@ -4789,7 +4789,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":844
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":844
  *             elif t == NPY_UINT:        f[0] =  73 #"I"
  *             elif t == NPY_LONG:        f[0] = 108 #"l"
  *             elif t == NPY_ULONG:       f[0] = 76  #"L"             # <<<<<<<<<<<<<<
@@ -4807,7 +4807,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":845
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":845
  *             elif t == NPY_LONG:        f[0] = 108 #"l"
  *             elif t == NPY_ULONG:       f[0] = 76  #"L"
  *             elif t == NPY_LONGLONG:    f[0] = 113 #"q"             # <<<<<<<<<<<<<<
@@ -4825,7 +4825,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":846
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":846
  *             elif t == NPY_ULONG:       f[0] = 76  #"L"
  *             elif t == NPY_LONGLONG:    f[0] = 113 #"q"
  *             elif t == NPY_ULONGLONG:   f[0] = 81  #"Q"             # <<<<<<<<<<<<<<
@@ -4843,7 +4843,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":847
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":847
  *             elif t == NPY_LONGLONG:    f[0] = 113 #"q"
  *             elif t == NPY_ULONGLONG:   f[0] = 81  #"Q"
  *             elif t == NPY_FLOAT:       f[0] = 102 #"f"             # <<<<<<<<<<<<<<
@@ -4861,7 +4861,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":848
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":848
  *             elif t == NPY_ULONGLONG:   f[0] = 81  #"Q"
  *             elif t == NPY_FLOAT:       f[0] = 102 #"f"
  *             elif t == NPY_DOUBLE:      f[0] = 100 #"d"             # <<<<<<<<<<<<<<
@@ -4879,7 +4879,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":849
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":849
  *             elif t == NPY_FLOAT:       f[0] = 102 #"f"
  *             elif t == NPY_DOUBLE:      f[0] = 100 #"d"
  *             elif t == NPY_LONGDOUBLE:  f[0] = 103 #"g"             # <<<<<<<<<<<<<<
@@ -4897,7 +4897,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":850
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":850
  *             elif t == NPY_DOUBLE:      f[0] = 100 #"d"
  *             elif t == NPY_LONGDOUBLE:  f[0] = 103 #"g"
  *             elif t == NPY_CFLOAT:      f[0] = 90; f[1] = 102; f += 1 # Zf             # <<<<<<<<<<<<<<
@@ -4917,7 +4917,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":851
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":851
  *             elif t == NPY_LONGDOUBLE:  f[0] = 103 #"g"
  *             elif t == NPY_CFLOAT:      f[0] = 90; f[1] = 102; f += 1 # Zf
  *             elif t == NPY_CDOUBLE:     f[0] = 90; f[1] = 100; f += 1 # Zd             # <<<<<<<<<<<<<<
@@ -4937,7 +4937,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":852
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":852
  *             elif t == NPY_CFLOAT:      f[0] = 90; f[1] = 102; f += 1 # Zf
  *             elif t == NPY_CDOUBLE:     f[0] = 90; f[1] = 100; f += 1 # Zd
  *             elif t == NPY_CLONGDOUBLE: f[0] = 90; f[1] = 103; f += 1 # Zg             # <<<<<<<<<<<<<<
@@ -4957,7 +4957,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":853
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":853
  *             elif t == NPY_CDOUBLE:     f[0] = 90; f[1] = 100; f += 1 # Zd
  *             elif t == NPY_CLONGDOUBLE: f[0] = 90; f[1] = 103; f += 1 # Zg
  *             elif t == NPY_OBJECT:      f[0] = 79 #"O"             # <<<<<<<<<<<<<<
@@ -4975,7 +4975,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
         goto __pyx_L15;
       }
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":855
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":855
  *             elif t == NPY_OBJECT:      f[0] = 79 #"O"
  *             else:
  *                 raise ValueError(u"unknown dtype code in numpy.pxd (%d)" % t)             # <<<<<<<<<<<<<<
@@ -4994,7 +4994,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       }
       __pyx_L15:;
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":856
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":856
  *             else:
  *                 raise ValueError(u"unknown dtype code in numpy.pxd (%d)" % t)
  *             f += 1             # <<<<<<<<<<<<<<
@@ -5003,7 +5003,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
  */
       __pyx_v_f = (__pyx_v_f + 1);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":831
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":831
  *         offset[0] += child.itemsize
  * 
  *         if not PyDataType_HASFIELDS(child):             # <<<<<<<<<<<<<<
@@ -5013,7 +5013,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
       goto __pyx_L13;
     }
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":860
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":860
  *             # Cython ignores struct boundary information ("T{...}"),
  *             # so don't output it
  *             f = _util_dtypestring(child, f, end, offset)             # <<<<<<<<<<<<<<
@@ -5026,7 +5026,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
     }
     __pyx_L13:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":805
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":805
  *     cdef tuple fields
  * 
  *     for childname in descr.names:             # <<<<<<<<<<<<<<
@@ -5036,7 +5036,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
   }
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":861
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":861
  *             # so don't output it
  *             f = _util_dtypestring(child, f, end, offset)
  *     return f             # <<<<<<<<<<<<<<
@@ -5046,7 +5046,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
   __pyx_r = __pyx_v_f;
   goto __pyx_L0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":796
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":796
  *         return ()
  * 
  * cdef inline char* _util_dtypestring(dtype descr, char* f, char* end, int* offset) except NULL:             # <<<<<<<<<<<<<<
@@ -5071,7 +5071,7 @@ static CYTHON_INLINE char *__pyx_f_5numpy__util_dtypestring(PyArray_Descr *__pyx
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":977
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":977
  * 
  * 
  * cdef inline void set_array_base(ndarray arr, object base):             # <<<<<<<<<<<<<<
@@ -5086,7 +5086,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
   int __pyx_t_2;
   __Pyx_RefNannySetupContext("set_array_base", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":979
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":979
  * cdef inline void set_array_base(ndarray arr, object base):
  *      cdef PyObject* baseptr
  *      if base is None:             # <<<<<<<<<<<<<<
@@ -5097,7 +5097,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
   __pyx_t_2 = (__pyx_t_1 != 0);
   if (__pyx_t_2) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":980
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":980
  *      cdef PyObject* baseptr
  *      if base is None:
  *          baseptr = NULL             # <<<<<<<<<<<<<<
@@ -5106,7 +5106,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
  */
     __pyx_v_baseptr = NULL;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":979
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":979
  * cdef inline void set_array_base(ndarray arr, object base):
  *      cdef PyObject* baseptr
  *      if base is None:             # <<<<<<<<<<<<<<
@@ -5116,7 +5116,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
     goto __pyx_L3;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":982
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":982
  *          baseptr = NULL
  *      else:
  *          Py_INCREF(base) # important to do this before decref below!             # <<<<<<<<<<<<<<
@@ -5126,7 +5126,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
   /*else*/ {
     Py_INCREF(__pyx_v_base);
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":983
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":983
  *      else:
  *          Py_INCREF(base) # important to do this before decref below!
  *          baseptr = <PyObject*>base             # <<<<<<<<<<<<<<
@@ -5137,7 +5137,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
   }
   __pyx_L3:;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":984
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":984
  *          Py_INCREF(base) # important to do this before decref below!
  *          baseptr = <PyObject*>base
  *      Py_XDECREF(arr.base)             # <<<<<<<<<<<<<<
@@ -5146,7 +5146,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
  */
   Py_XDECREF(__pyx_v_arr->base);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":985
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":985
  *          baseptr = <PyObject*>base
  *      Py_XDECREF(arr.base)
  *      arr.base = baseptr             # <<<<<<<<<<<<<<
@@ -5155,7 +5155,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
  */
   __pyx_v_arr->base = __pyx_v_baseptr;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":977
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":977
  * 
  * 
  * cdef inline void set_array_base(ndarray arr, object base):             # <<<<<<<<<<<<<<
@@ -5167,7 +5167,7 @@ static CYTHON_INLINE void __pyx_f_5numpy_set_array_base(PyArrayObject *__pyx_v_a
   __Pyx_RefNannyFinishContext();
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":987
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":987
  *      arr.base = baseptr
  * 
  * cdef inline object get_array_base(ndarray arr):             # <<<<<<<<<<<<<<
@@ -5181,7 +5181,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
   int __pyx_t_1;
   __Pyx_RefNannySetupContext("get_array_base", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":988
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":988
  * 
  * cdef inline object get_array_base(ndarray arr):
  *     if arr.base is NULL:             # <<<<<<<<<<<<<<
@@ -5191,7 +5191,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
   __pyx_t_1 = ((__pyx_v_arr->base == NULL) != 0);
   if (__pyx_t_1) {
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":989
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":989
  * cdef inline object get_array_base(ndarray arr):
  *     if arr.base is NULL:
  *         return None             # <<<<<<<<<<<<<<
@@ -5202,7 +5202,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
     __pyx_r = Py_None; __Pyx_INCREF(Py_None);
     goto __pyx_L0;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":988
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":988
  * 
  * cdef inline object get_array_base(ndarray arr):
  *     if arr.base is NULL:             # <<<<<<<<<<<<<<
@@ -5211,7 +5211,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
  */
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":991
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":991
  *         return None
  *     else:
  *         return <object>arr.base             # <<<<<<<<<<<<<<
@@ -5225,7 +5225,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
     goto __pyx_L0;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":987
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":987
  *      arr.base = baseptr
  * 
  * cdef inline object get_array_base(ndarray arr):             # <<<<<<<<<<<<<<
@@ -5240,7 +5240,7 @@ static CYTHON_INLINE PyObject *__pyx_f_5numpy_get_array_base(PyArrayObject *__py
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":996
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":996
  * # Versions of the import_* functions which are more suitable for
  * # Cython code.
  * cdef inline int import_array() except -1:             # <<<<<<<<<<<<<<
@@ -5261,7 +5261,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
   PyObject *__pyx_t_8 = NULL;
   __Pyx_RefNannySetupContext("import_array", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
  * # Cython code.
  * cdef inline int import_array() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5277,7 +5277,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
     __Pyx_XGOTREF(__pyx_t_3);
     /*try:*/ {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":998
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":998
  * cdef inline int import_array() except -1:
  *     try:
  *         _import_array()             # <<<<<<<<<<<<<<
@@ -5286,7 +5286,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
  */
       __pyx_t_4 = _import_array(); if (unlikely(__pyx_t_4 == ((int)-1))) __PYX_ERR(1, 998, __pyx_L3_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
  * # Cython code.
  * cdef inline int import_array() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5300,7 +5300,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
     goto __pyx_L8_try_end;
     __pyx_L3_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":999
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":999
  *     try:
  *         _import_array()
  *     except Exception:             # <<<<<<<<<<<<<<
@@ -5315,7 +5315,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
       __Pyx_GOTREF(__pyx_t_6);
       __Pyx_GOTREF(__pyx_t_7);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1000
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1000
  *         _import_array()
  *     except Exception:
  *         raise ImportError("numpy.core.multiarray failed to import")             # <<<<<<<<<<<<<<
@@ -5331,7 +5331,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
     goto __pyx_L5_except_error;
     __pyx_L5_except_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":997
  * # Cython code.
  * cdef inline int import_array() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5346,7 +5346,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
     __pyx_L8_try_end:;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":996
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":996
  * # Versions of the import_* functions which are more suitable for
  * # Cython code.
  * cdef inline int import_array() except -1:             # <<<<<<<<<<<<<<
@@ -5369,7 +5369,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_array(void) {
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1002
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1002
  *         raise ImportError("numpy.core.multiarray failed to import")
  * 
  * cdef inline int import_umath() except -1:             # <<<<<<<<<<<<<<
@@ -5390,7 +5390,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
   PyObject *__pyx_t_8 = NULL;
   __Pyx_RefNannySetupContext("import_umath", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
  * 
  * cdef inline int import_umath() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5406,7 +5406,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
     __Pyx_XGOTREF(__pyx_t_3);
     /*try:*/ {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1004
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1004
  * cdef inline int import_umath() except -1:
  *     try:
  *         _import_umath()             # <<<<<<<<<<<<<<
@@ -5415,7 +5415,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
  */
       __pyx_t_4 = _import_umath(); if (unlikely(__pyx_t_4 == ((int)-1))) __PYX_ERR(1, 1004, __pyx_L3_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
  * 
  * cdef inline int import_umath() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5429,7 +5429,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
     goto __pyx_L8_try_end;
     __pyx_L3_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1005
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1005
  *     try:
  *         _import_umath()
  *     except Exception:             # <<<<<<<<<<<<<<
@@ -5444,7 +5444,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
       __Pyx_GOTREF(__pyx_t_6);
       __Pyx_GOTREF(__pyx_t_7);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1006
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1006
  *         _import_umath()
  *     except Exception:
  *         raise ImportError("numpy.core.umath failed to import")             # <<<<<<<<<<<<<<
@@ -5460,7 +5460,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
     goto __pyx_L5_except_error;
     __pyx_L5_except_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1003
  * 
  * cdef inline int import_umath() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5475,7 +5475,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
     __pyx_L8_try_end:;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1002
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1002
  *         raise ImportError("numpy.core.multiarray failed to import")
  * 
  * cdef inline int import_umath() except -1:             # <<<<<<<<<<<<<<
@@ -5498,7 +5498,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_umath(void) {
   return __pyx_r;
 }
 
-/* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
+/* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
  *         raise ImportError("numpy.core.umath failed to import")
  * 
  * cdef inline int import_ufunc() except -1:             # <<<<<<<<<<<<<<
@@ -5519,7 +5519,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
   PyObject *__pyx_t_8 = NULL;
   __Pyx_RefNannySetupContext("import_ufunc", 0);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
  * 
  * cdef inline int import_ufunc() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5535,7 +5535,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
     __Pyx_XGOTREF(__pyx_t_3);
     /*try:*/ {
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1010
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1010
  * cdef inline int import_ufunc() except -1:
  *     try:
  *         _import_umath()             # <<<<<<<<<<<<<<
@@ -5544,7 +5544,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
  */
       __pyx_t_4 = _import_umath(); if (unlikely(__pyx_t_4 == ((int)-1))) __PYX_ERR(1, 1010, __pyx_L3_error)
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
  * 
  * cdef inline int import_ufunc() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5558,7 +5558,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
     goto __pyx_L8_try_end;
     __pyx_L3_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1011
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1011
  *     try:
  *         _import_umath()
  *     except Exception:             # <<<<<<<<<<<<<<
@@ -5572,7 +5572,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
       __Pyx_GOTREF(__pyx_t_6);
       __Pyx_GOTREF(__pyx_t_7);
 
-      /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1012
+      /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1012
  *         _import_umath()
  *     except Exception:
  *         raise ImportError("numpy.core.umath failed to import")             # <<<<<<<<<<<<<<
@@ -5586,7 +5586,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
     goto __pyx_L5_except_error;
     __pyx_L5_except_error:;
 
-    /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
+    /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1009
  * 
  * cdef inline int import_ufunc() except -1:
  *     try:             # <<<<<<<<<<<<<<
@@ -5601,7 +5601,7 @@ static CYTHON_INLINE int __pyx_f_5numpy_import_ufunc(void) {
     __pyx_L8_try_end:;
   }
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
  *         raise ImportError("numpy.core.umath failed to import")
  * 
  * cdef inline int import_ufunc() except -1:             # <<<<<<<<<<<<<<
@@ -5709,35 +5709,35 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_RefNannyDeclarations
   __Pyx_RefNannySetupContext("__Pyx_InitCachedConstants", 0);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":70
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":68
  * 
  * cdef pmatrix_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)             # <<<<<<<<<<<<<<
  * 
  *     mat[0, 0] = p11_c(iza, raa, vza)
  */
-  __pyx_tuple_ = PyTuple_Pack(2, __pyx_int_4, __pyx_int_4); if (unlikely(!__pyx_tuple_)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_tuple_ = PyTuple_Pack(2, __pyx_int_4, __pyx_int_4); if (unlikely(!__pyx_tuple_)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple_);
   __Pyx_GIVEREF(__pyx_tuple_);
-  __pyx_tuple__2 = PyTuple_Pack(1, __pyx_tuple_); if (unlikely(!__pyx_tuple__2)) __PYX_ERR(0, 70, __pyx_L1_error)
+  __pyx_tuple__2 = PyTuple_Pack(1, __pyx_tuple_); if (unlikely(!__pyx_tuple__2)) __PYX_ERR(0, 68, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple__2);
   __Pyx_GIVEREF(__pyx_tuple__2);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":90
- * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":87
  * 
+ * cdef dblquad_c(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):
  *     cdef np.ndarray[DTYPE_t, ndim=2] mat = np.zeros((4, 4), dtype=np.float)             # <<<<<<<<<<<<<<
  * 
  *     mat[0, 0] = pi * sin(vza) ** 2 + pi / 3
  */
-  __pyx_tuple__3 = PyTuple_Pack(2, __pyx_int_4, __pyx_int_4); if (unlikely(!__pyx_tuple__3)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_tuple__3 = PyTuple_Pack(2, __pyx_int_4, __pyx_int_4); if (unlikely(!__pyx_tuple__3)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple__3);
   __Pyx_GIVEREF(__pyx_tuple__3);
-  __pyx_tuple__4 = PyTuple_Pack(1, __pyx_tuple__3); if (unlikely(!__pyx_tuple__4)) __PYX_ERR(0, 90, __pyx_L1_error)
+  __pyx_tuple__4 = PyTuple_Pack(1, __pyx_tuple__3); if (unlikely(!__pyx_tuple__4)) __PYX_ERR(0, 87, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple__4);
   __Pyx_GIVEREF(__pyx_tuple__4);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":229
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":229
  *             if ((flags & pybuf.PyBUF_C_CONTIGUOUS == pybuf.PyBUF_C_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_C_CONTIGUOUS)):
  *                 raise ValueError(u"ndarray is not C contiguous")             # <<<<<<<<<<<<<<
@@ -5748,7 +5748,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__5);
   __Pyx_GIVEREF(__pyx_tuple__5);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":233
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":233
  *             if ((flags & pybuf.PyBUF_F_CONTIGUOUS == pybuf.PyBUF_F_CONTIGUOUS)
  *                 and not PyArray_CHKFLAGS(self, NPY_F_CONTIGUOUS)):
  *                 raise ValueError(u"ndarray is not Fortran contiguous")             # <<<<<<<<<<<<<<
@@ -5759,7 +5759,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__6);
   __Pyx_GIVEREF(__pyx_tuple__6);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":263
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":263
  *                 if ((descr.byteorder == c'>' and little_endian) or
  *                     (descr.byteorder == c'<' and not little_endian)):
  *                     raise ValueError(u"Non-native byte order not supported")             # <<<<<<<<<<<<<<
@@ -5770,7 +5770,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__7);
   __Pyx_GIVEREF(__pyx_tuple__7);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":810
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":810
  * 
  *         if (end - f) - <int>(new_offset - offset[0]) < 15:
  *             raise RuntimeError(u"Format string allocated too short, see comment in numpy.pxd")             # <<<<<<<<<<<<<<
@@ -5781,7 +5781,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__8);
   __Pyx_GIVEREF(__pyx_tuple__8);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":814
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":814
  *         if ((child.byteorder == c'>' and little_endian) or
  *             (child.byteorder == c'<' and not little_endian)):
  *             raise ValueError(u"Non-native byte order not supported")             # <<<<<<<<<<<<<<
@@ -5792,7 +5792,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__9);
   __Pyx_GIVEREF(__pyx_tuple__9);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":834
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":834
  *             t = child.type_num
  *             if end - f < 5:
  *                 raise RuntimeError(u"Format string allocated too short.")             # <<<<<<<<<<<<<<
@@ -5803,7 +5803,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__10);
   __Pyx_GIVEREF(__pyx_tuple__10);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1000
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1000
  *         _import_array()
  *     except Exception:
  *         raise ImportError("numpy.core.multiarray failed to import")             # <<<<<<<<<<<<<<
@@ -5814,7 +5814,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__11);
   __Pyx_GIVEREF(__pyx_tuple__11);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1006
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1006
  *         _import_umath()
  *     except Exception:
  *         raise ImportError("numpy.core.umath failed to import")             # <<<<<<<<<<<<<<
@@ -5825,7 +5825,7 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__12);
   __Pyx_GIVEREF(__pyx_tuple__12);
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1012
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1012
  *         _import_umath()
  *     except Exception:
  *         raise ImportError("numpy.core.umath failed to import")             # <<<<<<<<<<<<<<
@@ -5834,28 +5834,28 @@ static int __Pyx_InitCachedConstants(void) {
   __Pyx_GOTREF(__pyx_tuple__13);
   __Pyx_GIVEREF(__pyx_tuple__13);
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":97
  *     return mat
  * 
  * def dblquad_c_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
  *     return dblquad_c(iza, vza, raa)
  * 
  */
-  __pyx_tuple__14 = PyTuple_Pack(3, __pyx_n_s_iza, __pyx_n_s_vza, __pyx_n_s_raa); if (unlikely(!__pyx_tuple__14)) __PYX_ERR(0, 100, __pyx_L1_error)
+  __pyx_tuple__14 = PyTuple_Pack(3, __pyx_n_s_iza, __pyx_n_s_vza, __pyx_n_s_raa); if (unlikely(!__pyx_tuple__14)) __PYX_ERR(0, 97, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple__14);
   __Pyx_GIVEREF(__pyx_tuple__14);
-  __pyx_codeobj__15 = (PyObject*)__Pyx_PyCode_New(3, 0, 3, 0, CO_OPTIMIZED|CO_NEWLOCALS, __pyx_empty_bytes, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_tuple__14, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_kp_s_pyrism_scattering_rayleigh_core, __pyx_n_s_dblquad_c_wrapper, 100, __pyx_empty_bytes); if (unlikely(!__pyx_codeobj__15)) __PYX_ERR(0, 100, __pyx_L1_error)
+  __pyx_codeobj__15 = (PyObject*)__Pyx_PyCode_New(3, 0, 3, 0, CO_OPTIMIZED|CO_NEWLOCALS, __pyx_empty_bytes, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_tuple__14, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_kp_s_pyrism_scattering_rayleigh_core, __pyx_n_s_dblquad_c_wrapper, 97, __pyx_empty_bytes); if (unlikely(!__pyx_codeobj__15)) __PYX_ERR(0, 97, __pyx_L1_error)
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":103
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
  *     return dblquad_c(iza, vza, raa)
  * 
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- *     return pmatrix_c( iza, vza, raa)
+ *     return pmatrix_c(iza, vza, raa)
  */
-  __pyx_tuple__16 = PyTuple_Pack(3, __pyx_n_s_iza, __pyx_n_s_vza, __pyx_n_s_raa); if (unlikely(!__pyx_tuple__16)) __PYX_ERR(0, 103, __pyx_L1_error)
+  __pyx_tuple__16 = PyTuple_Pack(3, __pyx_n_s_iza, __pyx_n_s_vza, __pyx_n_s_raa); if (unlikely(!__pyx_tuple__16)) __PYX_ERR(0, 100, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_tuple__16);
   __Pyx_GIVEREF(__pyx_tuple__16);
-  __pyx_codeobj__17 = (PyObject*)__Pyx_PyCode_New(3, 0, 3, 0, CO_OPTIMIZED|CO_NEWLOCALS, __pyx_empty_bytes, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_tuple__16, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_kp_s_pyrism_scattering_rayleigh_core, __pyx_n_s_pmatrix_wrapper, 103, __pyx_empty_bytes); if (unlikely(!__pyx_codeobj__17)) __PYX_ERR(0, 103, __pyx_L1_error)
+  __pyx_codeobj__17 = (PyObject*)__Pyx_PyCode_New(3, 0, 3, 0, CO_OPTIMIZED|CO_NEWLOCALS, __pyx_empty_bytes, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_tuple__16, __pyx_empty_tuple, __pyx_empty_tuple, __pyx_kp_s_pyrism_scattering_rayleigh_core, __pyx_n_s_pmatrix_wrapper, 100, __pyx_empty_bytes); if (unlikely(!__pyx_codeobj__17)) __PYX_ERR(0, 100, __pyx_L1_error)
   __Pyx_RefNannyFinishContext();
   return 0;
   __pyx_L1_error:;
@@ -6141,7 +6141,7 @@ if (!__Pyx_RefNanny) {
  * from libc.math cimport cos, sin, pow, pi
  * from scipy.integrate import dblquad as sdblquad             # <<<<<<<<<<<<<<
  * 
- * 
+ * DTYPE = np.float
  */
   __pyx_t_1 = PyList_New(1); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 6, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
@@ -6157,42 +6157,42 @@ if (!__Pyx_RefNanny) {
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
   __Pyx_DECREF(__pyx_t_2); __pyx_t_2 = 0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":9
- * 
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":8
+ * from scipy.integrate import dblquad as sdblquad
  * 
  * DTYPE = np.float             # <<<<<<<<<<<<<<
  * ctypedef np.float_t DTYPE_t
  * 
  */
-  __pyx_t_2 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 9, __pyx_L1_error)
+  __pyx_t_2 = __Pyx_GetModuleGlobalName(__pyx_n_s_np); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 8, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_2);
-  __pyx_t_1 = __Pyx_PyObject_GetAttrStr(__pyx_t_2, __pyx_n_s_float); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 9, __pyx_L1_error)
+  __pyx_t_1 = __Pyx_PyObject_GetAttrStr(__pyx_t_2, __pyx_n_s_float); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 8, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
   __Pyx_DECREF(__pyx_t_2); __pyx_t_2 = 0;
-  if (PyDict_SetItem(__pyx_d, __pyx_n_s_DTYPE, __pyx_t_1) < 0) __PYX_ERR(0, 9, __pyx_L1_error)
+  if (PyDict_SetItem(__pyx_d, __pyx_n_s_DTYPE, __pyx_t_1) < 0) __PYX_ERR(0, 8, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":97
  *     return mat
  * 
  * def dblquad_c_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
  *     return dblquad_c(iza, vza, raa)
  * 
  */
-  __pyx_t_1 = PyCFunction_NewEx(&__pyx_mdef_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_1dblquad_c_wrapper, NULL, __pyx_n_s_pyrism_scattering_rayleigh_core_2); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 100, __pyx_L1_error)
+  __pyx_t_1 = PyCFunction_NewEx(&__pyx_mdef_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_1dblquad_c_wrapper, NULL, __pyx_n_s_pyrism_scattering_rayleigh_core_2); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 97, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  if (PyDict_SetItem(__pyx_d, __pyx_n_s_dblquad_c_wrapper, __pyx_t_1) < 0) __PYX_ERR(0, 100, __pyx_L1_error)
+  if (PyDict_SetItem(__pyx_d, __pyx_n_s_dblquad_c_wrapper, __pyx_t_1) < 0) __PYX_ERR(0, 97, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
 
-  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":103
+  /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":100
  *     return dblquad_c(iza, vza, raa)
  * 
  * def pmatrix_wrapper(DTYPE_t iza, DTYPE_t vza, DTYPE_t raa):             # <<<<<<<<<<<<<<
- *     return pmatrix_c( iza, vza, raa)
+ *     return pmatrix_c(iza, vza, raa)
  */
-  __pyx_t_1 = PyCFunction_NewEx(&__pyx_mdef_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_3pmatrix_wrapper, NULL, __pyx_n_s_pyrism_scattering_rayleigh_core_2); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 103, __pyx_L1_error)
+  __pyx_t_1 = PyCFunction_NewEx(&__pyx_mdef_6pyrism_10scattering_8rayleigh_4core_16rayleigh_phase_c_3pmatrix_wrapper, NULL, __pyx_n_s_pyrism_scattering_rayleigh_core_2); if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 100, __pyx_L1_error)
   __Pyx_GOTREF(__pyx_t_1);
-  if (PyDict_SetItem(__pyx_d, __pyx_n_s_pmatrix_wrapper, __pyx_t_1) < 0) __PYX_ERR(0, 103, __pyx_L1_error)
+  if (PyDict_SetItem(__pyx_d, __pyx_n_s_pmatrix_wrapper, __pyx_t_1) < 0) __PYX_ERR(0, 100, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
 
   /* "pyrism/scattering/rayleigh/core/rayleigh_phase_c.pyx":1
@@ -6205,7 +6205,7 @@ if (!__Pyx_RefNanny) {
   if (PyDict_SetItem(__pyx_d, __pyx_n_s_test, __pyx_t_1) < 0) __PYX_ERR(0, 1, __pyx_L1_error)
   __Pyx_DECREF(__pyx_t_1); __pyx_t_1 = 0;
 
-  /* "../../../Anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
+  /* "../../../anaconda2/lib/site-packages/Cython/Includes/numpy/__init__.pxd":1008
  *         raise ImportError("numpy.core.umath failed to import")
  * 
  * cdef inline int import_ufunc() except -1:             # <<<<<<<<<<<<<<
